### 2024-06-23 #6923 @editorialbot
[PRE REVIEW]: Orbital Dynamics in X-ray stellar binary systems
**Submitting author:** @xragua (<a href="http://orcid.org/0009-0001-0378-7879">Graciela Sanjurjo Ferrín</a>)
**Repository:** https://github.com/xragua/xraybinaryorbit
**Branch with paper.md** (empty if default branch):
**Version:** v1
