+++ xraybinaryorbit/xraybinaryorbit.py
@@ -1170,13 +1170,15 @@ def absorption_column_through_orbit_theoretical(resolution=0.01,show_plot=True):
          
 # Ionization parameter map ###############################################################
 
+ # Ionization parameter map ###############################################################
+
 def ionization_map_phase(size_in_Rstar=3, min_color=None, max_color=None, save_plot=False, name="ionization_map"):
 
     print(    """
 --------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 
-    Generates a logarithmic ionization parameter map based on the stellar wind density
-    and the ionization parameters influenced by the source luminosity and distance to the compact object.
+    Generates a logarithmic ionization parameter map based on the stellar wind density, the luminosity and orbital parameters.
+    The uncolored area represents the X-ray shadow.
 
     Parameters:
     - size_in_Rstar (float, optional): Extent of the map from the stellar center in stellar radii. Default is 3.
@@ -1214,11 +1216,32 @@ def ionization_map_phase(size_in_Rstar=3, min_color=None, max_color=None, save_p
     Rorb_plot = (abar * (1 - eccentricity**2) / (1 + eccentricity * np.cos((th - periapsis / 360) * 2 * np.pi))) * Rstar_cm
     x = np.arange(1, size_in_Rstar, 0.01) * Rstar_cm
     v = vinf_cm_s * (1 - Rstar_cm / x)**beta
-    rho = M_dot_grams / (4 * np.pi * v * x**2)
+    ro = M_dot_grams / (4 * np.pi * v * x**2)
     
     # Calculate electron number density
     na = 6.02214076e23 / 1.00797
-    ne = rho * na
+    ne = ro * na
+    
+    #..............................................Angles for shadow
+
+    phase_ns = phase * 2 * np.pi
+    phase_ns_degrees = np.degrees(phase_ns)
+
+    alpha = np.arcsin(1/(Rorb/ Rstar_cm))
+    alpha_degrees = np.degrees(alpha)
+
+    alpha2 = np.arcsin(1/size_in_Rstar)
+    alpha_degrees2 = np.degrees(alpha2)
+
+    gamma_ = 180-(alpha_degrees2 + alpha_degrees)
+    rho_ = gamma_+phase_ns_degrees
+
+    rho = rho_*2*np.pi/360
+    gamma = gamma_*2*np.pi/360
+
+    phase_touch = (180-90-alpha_degrees)/360
+
+    #..............................................
     
     # Create a DataFrame to store chi results
     chi_result = pd.DataFrame(index=np.round(x / Rstar_cm, 3))
@@ -1234,6 +1257,7 @@ def ionization_map_phase(size_in_Rstar=3, min_color=None, max_color=None, save_p
     if not max_color:
         max_color = np.percentile(np.concatenate(chi_result.values), 90)
         print("max color coefficient is", round(max_color,2))  # Corrected print statement
+        
     if not min_color:
         min_color = np.percentile(np.concatenate(chi_result.values), 10)
         print("min color coefficient is", round(min_color,2))
@@ -1246,22 +1270,86 @@ def ionization_map_phase(size_in_Rstar=3, min_color=None, max_color=None, save_p
     
     # Calculate area between bounds and plot
     area_between_bounds = 0
-    for i in range(len(th)):
[…]
