+++ xraybinaryorbit/xraybinaryorbit.py
@@ -1170,9 +1170,8 @@ def absorption_column_through_orbit_theoretical(resolution=0.01,show_plot=True):
          
 # Ionization parameter map ###############################################################
 
- # Ionization parameter map ###############################################################
 
-def ionization_map_phase(size_in_Rstar=3, min_color=None, max_color=None, save_plot=False, name="ionization_map"):
+def ionization_map_phase(size_in_Rstar=0, min_color=None, max_color=None, save_plot=False, name="ionization_map"):
 
     print(    """
 --------------------------------------------------------------------------------------------------------------------------------------------------------------------------
@@ -1181,7 +1180,7 @@ def ionization_map_phase(size_in_Rstar=3, min_color=None, max_color=None, save_p
     The uncolored area represents the X-ray shadow.
 
     Parameters:
-    - size_in_Rstar (float, optional): Extent of the map from the stellar center in stellar radii. Default is 3.
+    - size_in_Rstar (float, optional): Extent of the map from the stellar center in stellar radii. Default is 2 times the semimajor.
     - min_color (float, optional): Minimum color scale value for the ionization parameter. Default is None.
     - max_color (float, optional): Maximum color scale value for the ionization parameter. Default is None.
     - save_plot (bool, optional): Boolean to decide if the plot should be saved. Default is False.
@@ -1214,6 +1213,11 @@ def ionization_map_phase(size_in_Rstar=3, min_color=None, max_color=None, save_p
     # Orbital radius calculations
     Rorb = (abar * (1 - eccentricity**2) / (1 + eccentricity * np.cos((phase - periapsis / 360) * 2 * np.pi))) * Rstar_cm
     Rorb_plot = (abar * (1 - eccentricity**2) / (1 + eccentricity * np.cos((th - periapsis / 360) * 2 * np.pi))) * Rstar_cm
+    
+    if size_in_Rstar==0:
+        size_in_Rstar = 2*max(Rorb_plot/Rstar_cm)
+    
+    
     x = np.arange(1, size_in_Rstar, 0.01) * Rstar_cm
     v = vinf_cm_s * (1 - Rstar_cm / x)**beta
     ro = M_dot_grams / (4 * np.pi * v * x**2)
@@ -1271,6 +1275,12 @@ def ionization_map_phase(size_in_Rstar=3, min_color=None, max_color=None, save_p
     # Calculate area between bounds and plot
     area_between_bounds = 0
     
+    bound_limit_1=[]
+    bound_limit_2=[]
+    bound_limit_3=[]
+    bound_limit_4=[]
+    phase_limit=[]
+    phase_limit2=[]
     #....................................................................................
     th_ = np.arange(phase_touch+phase ,phase + gamma_/360, 0.001)
     for i in range(len(th_)):
@@ -1293,9 +1303,23 @@ def ionization_map_phase(size_in_Rstar=3, min_color=None, max_color=None, save_p
             axs.scatter(np.tile(th_[i] * 2 * np.pi, len(x)), x / Rstar_cm, c=colors, cmap='rainbow')
 
             if len(x_chi_select) > 1:
-                axs.scatter(np.tile(th_[i] * 2 * np.pi, len(x_chi_select)), x_chi_select, color="k", alpha=0.3)
-                area_between_bounds += np.sum(x_chi_select * (x[1] - x[0]) * (th_[1] - th_[0]) * 2 * np.pi)
+                
[…]
