+++ xraybinaryorbit/xraybinaryorbit.py
@@ -20,6 +20,7 @@ from scipy.integrate import quad
 from matplotlib.colors import Normalize
 from matplotlib.cm import ScalarMappable
 import inspect
+import math
 
 warnings.filterwarnings("ignore")
 warnings.filterwarnings("ignore", category=RuntimeWarning)
@@ -164,6 +165,15 @@ def chi_squared_weighted(y_data, y_err, y_pred):
     y_pred = np.asarray(y_pred)
 
     chi_squared = np.sum(((y_data - y_pred) / y_err) ** 2)
+    return chi_squared
+    
+def chi_squared(y_data, y_pred):
+
+    y_data = np.asarray(y_data)
+    y_err = np.asarray(y_err)
+    y_pred = np.asarray(y_pred)
+
+    chi_squared = np.sum((y_data - y_pred) ** 2)
     return chi_squared
 
 #................................................... Easy way to manage values
@@ -505,7 +515,7 @@ def orbital_time_to_phase_(t,iphase, semimajor, orbitalperiod, eccentricity, per
     times_to_interpolate = t-min(t)
     
     phase_interpolator = interp1d(time, th, kind='cubic', fill_value="extrapolate")
-    phase = phase_interpolator(times_to_interpolate)+iphase
+    phase = phase_interpolator(times_to_interpolate)
     
     w_interpolator = interp1d(time, W_to_interpolate, kind='cubic', fill_value="extrapolate")
     W = w_interpolator(times_to_interpolate)
@@ -590,7 +600,7 @@ def rebin_bins_(t, x, sy, nbin):
     t_new=[]
     sc_new=[]
     
-    t, x, sy = preprocess_data(t, c, sc)
+    t, x, sy = preprocess_data(t, x, sy)
     
     for i in range(len(x)-nbin-1):
 
@@ -609,6 +619,7 @@ def rebin_bins_(t, x, sy, nbin):
             t_new.append(sum(t_bin*w)/sum(w))
 
     return t_new,c_new,sc_new
+
 ##########################################################################################
 ################################ THEORETICAL FUNCTIONS ###################################
 ##########################################################################################
@@ -1190,7 +1201,7 @@ def ionization_map_phase(size_in_Rstar=0, min_color=None, max_color=None, save_p
 
     Returns:
     - chi_result (pd.DataFrame): DataFrame containing the ionization parameter map.
-    - area (float): The calculated area between bounds.
+    - area (float): The calculated area between bounds in cm^2.
 
 --------------------------------------------------------------------------------------------------------------------------------------------------------------------------""")
 
@@ -1300,7 +1311,7 @@ def ionization_map_phase(size_in_Rstar=0, min_color=None, max_color=None, save_p
             colors = cmap(norm(np.log(chi)))
         
 
-            axs.scatter(np.tile(th_[i] * 2 * np.pi, len(x)), x / Rstar_cm, c=colors, cmap='rainbow')
+            axs.scatter(np.tile(th_[i] * 2 * np.pi, len(x)), x / Rstar_cm, c=colors, cmap='rainbow',alpha=0.3)
 
             if len(x_chi_select) > 1:
                 
@@ -1317,9 +1328,6 @@ def ionization_map_phase(size_in_Rstar=0, min_color=None, max_color=None, save_p
                     bound_limit_4.append(x_chi_select[idx4])
                     phase_limit2.append(th_[i])
                     
[…]
