### 2024-09-02 #6923 @warrickball
Hi @xragua, sorry for the delay on my side too. I'll start looking for reviewers by are typical means. If you can think of anyone from your research community or the [list of JOSS reviewers](https://reviewers.joss.theoj.org/) who you think would be a good reviewer (and doesn't represent a [conflict of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy), let me know.

While I'm doing that, though, I encourage you to have a pre-emptive look at the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html). Regarding the paper, note that the paper [shouldn't be much more than 1000 words](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain). It also [shouldn't contain API documentation](https://joss.readthedocs.io/en/latest/review_criteria.html#the-joss-paper), e.g., a list of functions, which I think would already save about 300 words.

I had a quick go at running the Jupyter Notebook in `example_fexxv_doppler_shifts/Fit_real_line_shift_disc.ipynb` and immediately found I had to install `pyswarm`, which I couldn't find mentioned in the documentation. We [expect installation instructions to clearly explain any dependencies](https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions). Python projects are usually provided as packages that can be installed with `pip`, though this can be directly from the repo with e.g.
```
$ pip install git+https://github.com/xragua/xraybinaryorbit
```
and doesn't have to be from PyPI. This also simplifies being able to import `xraybinaryorbit`.

I then got as far as the "bounds input form" but had no idea what values to use. So I'd also suggest you check that all the examples run from a fresh checkout. I'd suggest moving all the `examples_...` folders to `examples/...` to keep the root folder a bit tidier.

Ideally, you could use a documentation generation tool like [Sphinx](https://www.sphinx-doc.org) or [MkDocs](https://www.mkdocs.org/) and deploy docs to your personal GitHub page (e.g. `xragua.github.io/xraybinaryorbit`) or a third-party service like [Read the Docs](https://about.readthedocs.com/) but this isn't required. The API docs are usually then generated from docstrings. You _almost_ have docstrings, in that your functions print a help statement when they're executed. These could be turned into docstrings so that whatever interface a user uses—including a Jupyter Notebook—can query that information.
