### 2024-09-03 #6923 @xragua
Hi @warrickball, thank you for the detailed feedback. I'll begin making the suggested corrections right away.
In the meantime, I've identified a few potential reviewers who have a background closely aligned with my package's focus on astronomy, astrophysics, Python, and X-ray data analysis. These reviewers are:
@bonamem
@tddesjardins
@cmbiwer
However, I'm confident that any reviewer with expertise in both astrophysics and Python would be suitable.
Thanks again and best regards
