+++ paper/paper.md
@@ -32,13 +32,14 @@ bibliography: paper.bib
 
 ---
 
+
 # Summary
 
 X-ray astronomy is a young discipline, spanning no more than a few decades. The Earth's atmosphere is opaque to this type of radiation, so observations in this part of the spectrum had to wait until the beginning of the space era, with rocket launchers carrying X-ray telescopes to reveal the universe from a brand-new point of view.
 
 X-ray binary systems consist of two stars in close orbit around each other, where one of the stars is typically a compact object such as a neutron star or a black hole. The compact object accretes matter from its companion star, which can be a main sequence star, a giant star, or even another compact object. The X-ray radiation in these systems is generated through the accretion of matter from the companion's powerful stellar wind, typical of these early-type stars. Close binaries may become compact-object mergers and eventually sources of gravitational waves and/or short gamma-ray bursts. They will also provide insight into the behavior of matter under extreme gravitational and magnetic fields. Understanding these processes is fundamental to modern astrophysics and has driven numerous theoretical and observational studies.
 
-The xraybinaryoorbit package helps to unveil X-Ray binary orbital dynamics based on key theories including the conservation of angular momentum in orbital mechanics, the CAK model for radiation-driven stellar winds, accretion luminosity, ionization parameters, and the Doppler effect. Additionally, it provides tools for the Lomb-Scargle periodogram to identify periodic signals in unevenly sampled data.
+The xraybinaryoorbit package helps to unveil X-Ray binary orbital dynamics based on key theories including the conservation of angular momentum in orbital mechanics, the CAK model for radiation-driven stellar winds, accretion luminosity, ionization parameters, and the Doppler effect. 
 
 
 # Science behind
@@ -88,20 +89,6 @@ In astronomy, the Doppler effect is used to analyze the motion of celestial obje
 
 In the context of X-ray binaries, the Doppler effect is evident in the pulsations of a neutron star (NS) orbiting its companion, allowing precise determination of orbital parameters like radius, mass, inclination, and eccentricity. Additionally, the Doppler effect influences emission line energies when the emitting plasma is in motion.
 
-### Simple Models
-- **Conic Orbit**: representing the movement of the compact object around the companion.
-- **Logarithmic Spiral**: gravitational capture of matter which will eventually be captured by the stellar object.
-
-### Composed Models
-- **Disc in a Conic Orbit**: representing an accretion disc or ballistic movement around the compact object traveling around the companion.
-- **Logarithmic Spiral in a Conic Orbit**: gravitational capture of matter by the compact object traveling in a conic orbit around the companion.
-
-$$ r_{\text{main}} = \frac{b_{\text{main}} \cdot (1 - e_{\text{main}}^2)}{1 + e_{\text{main}} \cdot \cos(\phi_{\text{main}} - W_{\text{main}})} $$
-
-$$ r_{\text{secondary}} = \frac{b_{\text{secondary}} \cdot (1 - e_{\text{secondary}}^2)}{1 + e_{\text{secondary}} \cdot \cos(\phi_{\text{secondary}} - W_{\text{secondary}})} $$
-
-$$ r_{\text{spiral}} = a_{\text{spiral}} \times \exp(b \cdot 2\pi \phi) $$
-
 The general equation for the Doppler velocity in terms of the orbital phase is:
 
 $$ v_{D} = (-r\omega \sin\phi \sin i) $$
@@ -109,47 +96,7 @@ $$ v_{D} = (-r\omega \sin\phi \sin i) $$
 $$ \lambda_{D} = \lambda_{\text{rest}}\left(1+\frac{v_{D}}{c}\right) $$
 
[…]
