+++ setup.py
@@ -10,7 +10,7 @@ setup(
     author_email='graciela.sanjurjo@ua.es',  # Your email address
     url='https://github.com/xragua/xraybinaryorbit/',  # URL of your GitHub repository
     download_url='https://github.com/xragua/xraybinaryorbit/archive/refs/tags/one.tar.gz',  # Download URL for the package
-    keywords=['astronomy', 'x-ray, 'binary','orbit'],  # Keywords describing your package
+    keywords=['astronomy', 'x-ray', 'binary','orbit'],  # Keywords describing your package
     install_requires=[  # List of dependencies
         'numpy',
         'pandas',
@@ -20,17 +20,10 @@ setup(
         'astropy',
     ],
     classifiers=[
-        'Development Status :: v_01',  # Current development status
-        'Intended Audience :: Astronomers',  # Audience type
-        'Topic :: X-ray astronomy orbital modulations:: Analysis Tools',
-        'License :: OSI Approved :: MIT License',  # License
-        'Programming Language :: Python :: 3',  # Supported Python versions
-        'Programming Language :: Python :: 3.4',
-        'Programming Language :: Python :: 3.5',
-        'Programming Language :: Python :: 3.6',
-        'Programming Language :: Python :: 3.7',
-        'Programming Language :: Python :: 3.8',
-        'Programming Language :: Python :: 3.9',
+        'Development Status :: 1 - Planning',
+        'Programming Language :: Python :: 3',  # Current development status
+
     ],
 )
+
 
