### 2024-09-07 6f81ae1
Updated docstrings and incremented version to X.X.X
