+++ xraybinaryorbit.egg-info/requires.txt
@@ -1,0 +1,6 @@
+numpy
+pandas
+scipy
+pyswarm
+matplotlib
+astropy
