+++ docs/build/html/_sources/api.rst.txt
@@ -1,0 +1,9 @@
+.. xraybinaryorbit documentation master file, created by
+   sphinx-quickstart on Mon Sep  9 18:12:16 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+
+API Reference
+=============
+
