### 2024-09-10 cbeb7af
Updated to version 0.2.1 and prepared for PyPI upload
