+++ xraybinaryorbit/xraybinaryorbit.py
@@ -633,9 +633,8 @@ def rebin_bins_(t, x, sy, nbin):
 
 # CONIC ORBIT #############################################################################
 def doppler_orbit_theoretical(t, units="keV", show_plot=False, precision_for_phase=0.01):
-
     """
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+
     Computes the Doppler variation expected from orbital movement given a time array in seconds.
 
     Parameters:
@@ -648,9 +647,9 @@ def doppler_orbit_theoretical(t, units="keV", show_plot=False, precision_for_pha
     - t (array-like): Input time array.
     - x (array-like): Orbital phase array.
     - equation (array-like): Expected Doppler variation.
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-"""
-    parameter_names=["iphase", "semimajor", "orbitalperiod", "eccentricity", "periapsis", "inclination", "Rstar", "Mstar1","Mstar2", "wind_vel", "feature"]
+
+    """
+    r_names=["iphase", "semimajor", "orbitalperiod", "eccentricity", "periapsis", "inclination", "Rstar", "Mstar1","Mstar2", "wind_vel", "feature"]
     
     fixed_values = manage_parameters(parameter_names, "orbit")
     iphase, semimajor, orbitalperiod, eccentricity, periapsis, inclination, Rstar, Mstar1, Mstar2, wind_vel, feature = fixed_values
@@ -715,10 +714,8 @@ def doppler_orbit_theoretical(t, units="keV", show_plot=False, precision_for_pha
     return t,x,equation
     
 # SPIRAL #########################################################################################
-def doppler_spiral_theoretical(t,  units="keV", show_plot=False):
-
+def doppler_spiral_theoretical(t, units="keV", show_plot=False):
     """
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------
     Computes the Doppler variation expected from a spiral movement given a time array in seconds.
     
     A logarithmic spiral is a type of spiral that grows in size by a constant factor with each turn. Its equation in polar coordinates is
@@ -740,8 +737,8 @@ def doppler_spiral_theoretical(t,  units="keV", show_plot=False):
     Returns:
     - x (array-like): Orbital phase array.
     - equation (array-like): Expected Doppler variation.
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-"""
+    """
+
 
     parameter_names=["iphase_spiral", "semimajor_spiral", "b", "omega", "inclination_spiral", "feature"]
     
@@ -804,10 +801,8 @@ def doppler_spiral_theoretical(t,  units="keV", show_plot=False):
     return t,x,equation
     
 # ORBIT IN ORBIT #####################################################################################
-def doppler_disc_theoretical(t, units="keV",show_plot=False):
-
+def doppler_disc_theoretical(t, units="keV", show_plot=False):
     """
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------
     Computes the Doppler variation expected from an orbital movement in a main orbit,
     assuming a ballistic movement of plasma around a compact object or the movement of a mass
     entering an accretion disc.
@@ -823,9 +818,7 @@ def doppler_disc_theoretical(t, units="keV",show_plot=False):
     - x (array-like): Orbital phase array for the first orbit.
     - x2 (array-like): Orbital phase array for the second orbit.
     - equation (array-like): Expected Doppler variation.
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-"""
-
+    """
[…]
