### 2024-09-12 #7220 @editorialbot
[REVIEW]: Orbital Dynamics in X-ray stellar binary systems
**Submitting author:** @xragua (<a href="http://orcid.org/0009-0001-0378-7879">Graciela Sanjurjo Ferrín</a>)
**Repository:** https://github.com/xragua/xraybinaryorbit
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.00
