### 2024-09-16 de6c72f
Remove unnecessary files and add .gitignore
