+++ build/lib/xraybinaryorbit/__init__.py
@@ -1,3 +1,0 @@
-# __init__.py
-
-from .xraybinaryorbit import *
