+++ xraybinaryorbit.egg-info/requires.txt
@@ -1,6 +1,0 @@
-numpy
-pandas
-scipy
-pyswarm
-matplotlib
-astropy
