### 2024-09-18 f1c9afa
fix error, print description in every function and correct error in periods
