+++ xraybinaryorbit/xraybinaryorbit.py
@@ -633,7 +633,7 @@ def rebin_bins_(t, x, sy, nbin):
 
 # CONIC ORBIT #############################################################################
 def doppler_orbit_theoretical(t, units="keV", show_plot=False, precision_for_phase=0.01):
-    """
+    print("""
 
     Computes the Doppler variation expected from orbital movement given a time array in seconds.
 
@@ -648,8 +648,8 @@ def doppler_orbit_theoretical(t, units="keV", show_plot=False, precision_for_pha
     - x (array-like): Orbital phase array.
     - equation (array-like): Expected Doppler variation.
 
-    """
-    r_names=["iphase", "semimajor", "orbitalperiod", "eccentricity", "periapsis", "inclination", "Rstar", "Mstar1","Mstar2", "wind_vel", "feature"]
+    """)
+    parameter_names=["iphase", "semimajor", "orbitalperiod", "eccentricity", "periapsis", "inclination", "Rstar", "Mstar1","Mstar2", "wind_vel", "feature"]
     
     fixed_values = manage_parameters(parameter_names, "orbit")
     iphase, semimajor, orbitalperiod, eccentricity, periapsis, inclination, Rstar, Mstar1, Mstar2, wind_vel, feature = fixed_values
@@ -715,7 +715,7 @@ def doppler_orbit_theoretical(t, units="keV", show_plot=False, precision_for_pha
     
 # SPIRAL #########################################################################################
 def doppler_spiral_theoretical(t, units="keV", show_plot=False):
-    """
+    print("""
     Computes the Doppler variation expected from a spiral movement given a time array in seconds.
     
     A logarithmic spiral is a type of spiral that grows in size by a constant factor with each turn. Its equation in polar coordinates is
@@ -737,7 +737,7 @@ def doppler_spiral_theoretical(t, units="keV", show_plot=False):
     Returns:
     - x (array-like): Orbital phase array.
     - equation (array-like): Expected Doppler variation.
-    """
+    """)
 
 
     parameter_names=["iphase_spiral", "semimajor_spiral", "b", "omega", "inclination_spiral", "feature"]
@@ -802,7 +802,7 @@ def doppler_spiral_theoretical(t, units="keV", show_plot=False):
     
 # ORBIT IN ORBIT #####################################################################################
 def doppler_disc_theoretical(t, units="keV", show_plot=False):
-    """
+    print("""
     Computes the Doppler variation expected from an orbital movement in a main orbit,
     assuming a ballistic movement of plasma around a compact object or the movement of a mass
     entering an accretion disc.
@@ -818,7 +818,7 @@ def doppler_disc_theoretical(t, units="keV", show_plot=False):
     - x (array-like): Orbital phase array for the first orbit.
     - x2 (array-like): Orbital phase array for the second orbit.
     - equation (array-like): Expected Doppler variation.
-    """
+    """)
     parameter_names=["iphase", "semimajor", "orbitalperiod", "eccentricity", "periapsis", "inclination", "Rstar", "Mstar1", "Mstar2", "iphase2", "semimajor2", "orbitalperiod2", "eccentricity2", "periapsis2", "inclination2",  "Mass3","wind_vel", "feature"]
     
     fixed_values = manage_parameters(parameter_names, "disc")
@@ -900,7 +900,7 @@ def doppler_disc_theoretical(t, units="keV", show_plot=False):
     
 # SPIRAL IN ORBIT ####################################################################################
 def doppler_spiral_in_orbit_theoretical(t, units="keV", show_plot=False):
-    """
+    print("""
     This function requires a time array in seconds and returns the time, orbital phase, and Doppler variation
     expected under the assumption of an orbital movement with a logarithmic spiral component.
 
[…]
