### 2024-09-20 1707132
Improve some examples and print again description when function is called
