+++ examples/example_fexxv_doppler_shifts_CenX3/bounds_disc_bounds.txt
@@ -1,0 +1,3 @@
+iphase,semimajor,orbitalperiod,eccentricity,periapsis,inclination,Rstar,Mstar1,Mstar2,iphase2,semimajor2,orbitalperiod2,eccentricity2,periapsis2,inclination2,Mass3,feature,wind_vel
+0.15,1.58,2.087,0.0,0.0,70.04,11.8,1.0,1.33,0.1,0.91,0.14,0.445,0.0,75.0,1e-60,6.63,0.0
+0.25,1.62,2.09,0.1,1e-70,76.7,12.2,22.77,1.45,0.5,0.92,0.16,0.46,90.0,77.0,1e-14,6.7,1e-70
