### 2024-09-30 #7220 @warrickball
Hi @matteobachetti & @tddesjardins, I was wondering if you'd made any progress with your reviews? We encourage reviewers to comment on whichever aspects they have a moment for at any time, rather than waiting to deliver a single, monolithic review. That way authors can start working on changes while the rest of the review proceeds.

You can also raise issues (or pull requests) in the software repo but try to mention `openjournals/joss-reviews#7220` so that the activity is linked here.
