### 2024-10-02 8f2d94d
Improve documentation of functions
