+++ xraybinaryorbit/xraybinaryorbit.py
@@ -638,22 +638,38 @@ def rebin_bins_(t, x, sy, nbin):
 
 # CONIC ORBIT #############################################################################
 def doppler_orbit_theoretical(t, units="keV", show_plot=False, precision_for_phase=0.01):
-    print("""
-
+    """
     Computes the Doppler variation expected from orbital movement given a time array in seconds.
 
-    Parameters:
-    - t (array-like): Time array in seconds.
-    - units (str, optional): Units for the output. Default is "keV". Can be "keV", "s", or "angstrom".
-    - show_plot (bool, optional): If True, a plot of the orbit and Doppler evolution is shown and saved. Default is False.
-    - precision_for_phase (float, optional): Precision for phase calculation. Default is 0.01.
+    Parameters
+    ----------
+    t : array-like
+        Time array in seconds.
+    units : str, optional
+        Units for the output Doppler variation. Default is "keV". Options include:
+        - "keV": Doppler variation in keV.
+        - "s": Doppler variation in seconds.
+        - "angstrom": Doppler variation in angstroms.
+    show_plot : bool, optional
+        If True, displays and saves a plot of the orbit and Doppler evolution. Default is False.
+    precision_for_phase : float, optional
+        Precision for the phase calculation. Default is 0.01.
 
-    Returns:
-    - t (array-like): Input time array.
-    - x (array-like): Orbital phase array.
-    - equation (array-like): Expected Doppler variation.
+    Returns
+    -------
+    t : array-like
+        The input time array.
+    x : array-like
+        Orbital phase array corresponding to the input time array.
+    equation : array-like
+        Expected Doppler variation computed for the given orbital movement.
 
-    """)
+    Notes
+    -----
+    This function calculates the theoretical Doppler variation from orbital motion, with the option to visualize
+    the orbital path and its corresponding Doppler shift.
+    """
+
     parameter_names=["iphase", "semimajor", "orbitalperiod", "eccentricity", "periapsis", "inclination", "Rstar", "Mstar1","Mstar2", "wind_vel", "feature"]
     
     fixed_values = manage_parameters(parameter_names, "orbit")
@@ -720,29 +736,44 @@ def doppler_orbit_theoretical(t, units="keV", show_plot=False, precision_for_pha
     
 # SPIRAL #########################################################################################
 def doppler_spiral_theoretical(t, units="keV", show_plot=False):
-    print("""
-    Computes the Doppler variation expected from a spiral movement given a time array in seconds.
-    
-    A logarithmic spiral is a type of spiral that grows in size by a constant factor with each turn. Its equation in polar coordinates is
-    r = a * e^(b * θ), where:
-
-    r is the distance from the origin (radius)
-    omega is the angle from a reference direction (usually the positive x-axis)
-    a is the scale factor that determines how quickly the spiral grows
-    b is the rate of rotation, controlling the tightness or looseness of the spiral
-
+    """
     Computes the Doppler variation expected from a spiral movement given a time array in seconds.
 
-    Parameters:
-    - t (array-like): Time array in seconds.
-    - units (str, optional): Units for the output. Default is "keV". Can be "keV", "s", or "angstrom".
-    - show_plot (bool, optional): If True, a plot of the spiral and Doppler evolution is shown and saved. Default is False.
-    - A form will appear to input the necessary orbital parameters. These parameters will be saved in a .txt file in the current directory and automatically loaded in subsequent runs. This avoids the need to re-enter all parameters; you can modify only those that require adjustment.
+    A logarithmic spiral is a type of spiral that grows by a constant factor with each turn. The spiral equation
+    in polar coordinates is r = a * e^(b * θ), where:
     
-    Returns:
-    - x (array-like): Orbital phase array.
[…]
