### 2024-10-03 c6f487f
improve documentation of functions
