### 2024-10-03 4fdc117
improve documentation of functions
