+++ orbit.txt
@@ -1,2 +1,0 @@
-iphase,semimajor,orbitalperiod,eccentricity,periapsis,inclination,Rstar,Mstar1,Mstar2,wind_vel,feature
-0.0,1.0,2.0,0.2,120.0,60.0,22.0,12.0,1.2,1000.0,5.0
