+++ disc.txt
@@ -1,2 +1,0 @@
-iphase,semimajor,orbitalperiod,eccentricity,periapsis,inclination,Rstar,Mstar1,Mstar2,iphase2,semimajor2,orbitalperiod2,eccentricity2,periapsis2,inclination2,Mass3,wind_vel,feature
-0.0,2.0,3.0,0.1,200.0,78.0,22.0,11.0,2.0,0.5,0.1,0.3,0.5,200.0,78.0,1.0,100.0,2.0
