+++ density_through_orbit.txt
@@ -1,2 +1,0 @@
-semimajor,orbitalperiod,eccentricity,periapsis,Rstar,Mstar1,Mstar2,wind_infinite_velocity,Mass_loss_rate,beta
-2.0,2.0,0.1,200.0,22.0,12.0,2.0,1000.0,1e-05,0.8
