+++ spiral_in_orbit.txt
@@ -1,2 +1,0 @@
-iphase,semimajor,orbitalperiod,eccentricity,periapsis,inclination,iphase_spiral,semimajor_spiral,b,omega,inclination_spiral,Rstar,Mstar1,Mstar2,wind_vel,feature
-0.0,2.0,0.2,0.2,200.0,89.0,0.0,0.01,0.001,0.0001,67.0,12.0,22.0,2.0,1000.0,3.0
