+++ time_phase.txt
@@ -1,2 +1,0 @@
-iphase,semimajor,orbitalperiod,eccentricity,periapsis,Rstar,Mstar1,Mstar2
-0.0,2.0,2.0,0.2,120.0,22.0,12.0,2.0
