+++ .github/workflows/ci.yml
@@ -31,4 +31,4 @@ jobs:
 
     - name: Run tests
       run: |
-        pytest --cov=xraybinaryorbit
+        pytest 
