+++ .github/workflows/ci.yml
@@ -14,21 +14,23 @@ jobs:
 
     strategy:
       matrix:
-        python-version: [3.7, 3.8, 3.9, 3.10]
+        python-version: [3.7, 3.8, 3.9, 3.10, 3.11]  # Valid Python versions
 
     steps:
     - name: Checkout code
-      uses: actions/checkout@v2
+      uses: actions/checkout@v3  # Updated action
 
     - name: Set up Python
-      uses: actions/setup-python@v2
+      uses: actions/setup-python@v4  # Updated action
       with:
         python-version: ${{ matrix.python-version }}
 
     - name: Install dependencies
       run: |
         pip install pytest
+        pip install -r requirements.txt
 
     - name: Run tests
       run: |
-        pytest 
+        pytest --cov=xraybinaryorbit
+
