+++ .github/workflows/ci.yml
@@ -14,7 +14,7 @@ jobs:
 
     strategy:
       matrix:
-        python-version: [3.7, 3.8, 3.9, 3.10, 3.11]  # Valid Python versions
+        python-version: [3.8, 3.9]  # Valid Python versions
 
     steps:
     - name: Checkout code
