+++ .github/workflows/ci.yml
@@ -28,7 +28,7 @@ jobs:
     - name: Install dependencies
       run: |
         pip install pytest
-        pip install -r requirements.txt
+
 
     - name: Run tests
       run: |
