+++ .github/workflows/ci.yml
@@ -32,5 +32,5 @@ jobs:
 
     - name: Run tests
       run: |
-        pytest --cov=xraybinaryorbit
+        pytest 
 
