+++ .github/workflows/ci.yml
@@ -28,6 +28,7 @@ jobs:
     - name: Install dependencies
       run: |
         pip install pytest
+        pip install -r requirements.txt
 
 
     - name: Run tests
