+++ .github/workflows/ci.yml
@@ -29,6 +29,7 @@ jobs:
       run: |
         pip install pytest
         pip install -r requirements.txt
+        pip install .
 
 
     - name: Run tests
