+++ .github/workflows/ci.yml
@@ -29,7 +29,7 @@ jobs:
       run: |
         pip install pytest
         pip install -r requirements.txt
-        pip install .
+        pip install ./xraybinaryorbit
 
 
     - name: Run tests
