+++ .github/workflows/ci.yml
@@ -29,8 +29,7 @@ jobs:
       run: |
         pip install pytest
         pip install -r requirements.txt
-        pip install ./xraybinaryorbit
-
+        pip install .
 
     - name: Run tests
       run: |
