+++ requirements.txt
@@ -5,3 +5,4 @@ pyswarm
 matplotlib
 tk
 astropy
+monkeypatch
