+++ .github/workflows/ci.yml
@@ -27,7 +27,7 @@ jobs:
 
     - name: Install dependencies
       run: |
-        pip install pytest
+        pip install pytest>=6.0
         pip install -r requirements.txt
         pip install .
 
