+++ .github/workflows/ci.yml
@@ -29,7 +29,9 @@ jobs:
       run: |
         pip install pytest>=6.0
         pip install -r requirements.txt
-        pip install .
+        pip install git+https://github.com/xragua/xraybinaryorbit.git
+
+
 
     - name: Run tests
       run: |
