### 2024-10-04 08ccedf
Update test_absorption_column_through_orbit_theoretical.py
