+++ setup.py
@@ -1,17 +1,18 @@
-from distutils.core import setup
+from setuptools import setup, find_packages
 
 setup(
     name='xraybinaryorbit',  # Your package name
-    packages=['xraybinaryorbit'],  # The name of the package directory
-    version='0.2.3',  # Initial version
+    version='0.2.4',  # Version of your package
     license='MIT',  # License type
     description='A package for x-ray binary orbital modulations.',  # Short description
     author='LAEX',  # Your name
     author_email='graciela.sanjurjo@ua.es',  # Your email address
     url='https://github.com/xragua/xraybinaryorbit/',  # URL of your GitHub repository
     download_url='https://github.com/xragua/xraybinaryorbit/archive/refs/tags/one.tar.gz',  # Download URL for the package
-    keywords=['astronomy', 'x-ray', 'binary','orbit'],  # Keywords describing your package
-    install_requires=[  # List of dependencies
+    keywords=['astronomy', 'x-ray', 'binary', 'orbit'],  # Keywords describing your package
+
+    # Dependencies
+    install_requires=[
         'numpy',
         'pandas',
         'scipy',
@@ -19,11 +20,14 @@ setup(
         'matplotlib',
         'astropy',
     ],
+
+    packages=find_packages(),  # Automatically find packages inside your directory
+
     classifiers=[
         'Development Status :: 1 - Planning',
-        'Programming Language :: Python :: 3',  # Current development status
-
+        'Programming Language :: Python :: 3',
+        'License :: OSI Approved :: MIT License',
+        'Operating System :: OS Independent',
     ],
+    python_requires='>=3.6',  # Minimum Python version requirement
 )
-
-
