+++ .github/workflows/ci.yml
@@ -29,7 +29,7 @@ jobs:
       run: |
         pip install pytest>=6.0
         pip install -r requirements.txt
-        pip install git+https://github.com/xragua/xraybinaryorbit.git
+    
 
 
 
