+++ .github/workflows/ci.yml
@@ -27,11 +27,11 @@ jobs:
 
     - name: Install dependencies
       run: |
+        pip install --upgrade pip
+        pip install --upgrade xraybinaryorbit
         pip install pytest>=6.0
         pip install -r requirements.txt
-    
-
-
+  
 
     - name: Run tests
       run: |
