+++ xraybinaryorbit/__init__.py
@@ -1,3 +1,8 @@
 # __init__.py
 
+# Import everything from xraybinaryorbit (if needed)
 from .xraybinaryorbit import *
+
+# Explicitly expose _manage_parameters
+from .xraybinaryorbit import _manage_parameters, _orbital_phase_to_time, _orbital_time_to_phase
+
