+++ requirements.txt
@@ -1,4 +1,3 @@
-<<<<<<< HEAD
 numpy
 pandas
 scipy
@@ -8,506 +7,3 @@ tk
 astropy
 pytest>=6.0
 xraybinaryorbit
-=======
-absl-py==1.4.0
-aesara==2.9.3
-alabaster==0.7.16
-anaconda-client==1.11.1
-anaconda-navigator==2.4.0
-anaconda-project @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_6ddp1qj6b_/croots/recipe/anaconda-project_1660339893712/work
-annotated-types==0.7.0
-anyio @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/croot-t_zs64wy/anyio_1644482593257/work/dist
-appdirs==1.4.4
-applaunchservices @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_1fiu9w6r9i/croots/recipe/applaunchservices_1661854643906/work
-appnope @ file:///Users/ktietz/ci_310/appnope_1643965056645/work
-appscript @ file:///Users/ktietz/ci_310/appscript_1643965071544/work
-argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
-argon2-cffi-bindings @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/croot-wbf5edig/argon2-cffi-bindings_1644845754377/work
-arrow @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_cc82z4oqjq/croot/arrow_1676588146009/work
-arviz==0.19.0
-asdf==3.0.1
-asdf-astropy==0.5.0
-asdf-coordinates-schemas==0.2.0
-asdf-standard==1.0.3
-asdf-transform-schemas==0.4.0
-asdf-unit-schemas==0.1.0
-asdf-wcs-schemas==0.3.0
-astroid @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_e9yo8l0380/croot/astroid_1676904320145/work
-astropy==6.1.3
-astropy-iers-data==0.2024.9.2.0.33.23
-asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
-astunparse==1.6.3
-atomicwrites==1.4.0
-attrs @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_978y9aqcd7/croot/attrs_1668696180911/work
-Automat @ file:///tmp/build/80754af9/automat_1600298431173/work
-autopep8 @ file:///opt/conda/conda-bld/autopep8_1650463822033/work
-babel==2.16.0
-backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
-backports.functools-lru-cache @ file:///tmp/build/80754af9/backports.functools_lru_cache_1618170165463/work
-backports.tempfile @ file:///home/linux1/recipes/ci/backports.tempfile_1610991236607/work
-backports.weakref==1.0.post1
-bambi==0.14.0
-bayesian-optimization==1.5.1
-bcrypt @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_03v55xv0yj/croots/recipe/bcrypt_1659554335339/work
-beartype==0.16.4
[…]
