+++ test/test_doppler_spiral_in_orbit_theoretical.py
@@ -1,39 +1,33 @@
 import pytest
 import numpy as np
+import sys
+
+sys.path.append('/Users/graci/Desktop/git/xraybinaryorbit/xraybinaryorbit')
 from xraybinaryorbit import *
 
-# Mock functions for _manage_parameters and _orbital_time_to_phase
-def _mock_manage_parameters(parameter_names, context):
-    """Mock function to return dummy fixed values for the spiral in orbit parameters."""
-    return [0.1, 10, 100, 0.5, 30, 45, 0.2, 5, 0.05, 2 * np.pi, 45, 1.2, 1.5, 0.8, 500, 6.4]  # Example values for parameters
-
-def _mock_orbital_time_to_phase(t, iphase, semimajor, orbitalperiod, eccentricity, periapsis, Rstar, Mstar1, Mstar2, precision=0.01):
-    """Mock function to simulate phase calculation."""
-    x = np.linspace(0, 1, len(t))  # Simulate orbital phase values
-    W = 1.0  # A simple placeholder for W
-    return x, None, W
-
 # Test for doppler_spiral_in_orbit_theoretical with keV units
-def test_doppler_spiral_in_orbit_theoretical_keV(monkeypatch):
-    monkeypatch.setattr('xraybinaryorbit._manage_parameters', _mock_manage_parameters)
-    monkeypatch.setattr('xraybinaryorbit._orbital_time_to_phase', _mock_orbital_time_to_phase)
-
+def test_doppler_spiral_in_orbit_theoretical_keV():
+    # Example time array
     t = np.arange(0, 1000, 1)
-    t_out, x_out, x2_out, equation_out = doppler_spiral_in_orbit_theoretical(t, units="keV")
 
+    # Run the function with keV units and load parameters directly from the file
+    t_out, x_out, x2_out, equation_out = doppler_spiral_in_orbit_theoretical(t, units="keV", load_directly=True)
+
+    # Assert outputs
     assert len(t_out) == len(t), "The output time array should match the input length."
     assert len(x_out) == len(t), "The output orbital phase array should match the input length."
     assert len(x2_out) == len(t), "The second orbital phase array should match the input length."
     assert len(equation_out) == len(t), "The Doppler variation array should match the input length."
     assert isinstance(equation_out, np.ndarray), "Doppler variation should be a NumPy array."
 
-# Test for plot generation with keV units
-def test_doppler_spiral_in_orbit_theoretical_plot_keV(monkeypatch):
-    monkeypatch.setattr('xraybinaryorbit._manage_parameters', _mock_manage_parameters)
-    monkeypatch.setattr('xraybinaryorbit._orbital_time_to_phase', _mock_orbital_time_to_phase)
 
+# Test for plot generation with keV units
+def test_doppler_spiral_in_orbit_theoretical_plot_keV():
+    # Example time array
     t = np.arange(0, 1000, 1)
+
+    # Run the function with plot generation enabled (should not raise any exceptions)
     try:
-        doppler_spiral_in_orbit_theoretical(t, units="keV", show_plot=True)
+        doppler_spiral_in_orbit_theoretical(t, units="keV", show_plot=True, load_directly=True)
     except Exception as e:
         pytest.fail(f"Plot generation failed with error: {e}")
