+++ test/test_density_through_orbit_theoretical.py
@@ -1,55 +1,41 @@
 import pytest
 import numpy as np
+import sys
+
+sys.path.append('/Users/graci/Desktop/git/xraybinaryorbit/xraybinaryorbit')
 from xraybinaryorbit import *
 
-# Mock functions for _manage_parameters and _orbital_phase_to_time
-def _mock_manage_parameters(parameter_names, context):
-    """Mock function to return dummy fixed values for the density through orbit parameters."""
-    return [10, 100, 0.5, 30, 45, 1.5, 0.8, 1000, 2e-5, 0.8]  # Example values for parameters
-
-def _mock_orbital_phase_to_time(th, iphase, semimajor, orbitalperiod, eccentricity, periapsis, Rstar, Mstar1, Mstar2, precision=0.01):
-    """Mock function to simulate phase to time calculation."""
-    time = th * orbitalperiod * 24 * 3600  # Simulate time array in seconds
-    W = 1.0  # A simple placeholder for W
-    return time, time, W
-
-# Test for density_through_orbit_theoretical with default resolution
-def test_density_through_orbit_theoretical_default(monkeypatch):
-    # Patch the internal functions
-    monkeypatch.setattr('xraybinaryorbit._manage_parameters', _mock_manage_parameters)
-    monkeypatch.setattr('xraybinaryorbit._orbital_phase_to_time', _mock_orbital_phase_to_time)
+# Test for the default case with plot generation disabled
+def test_density_through_orbit_theoretical_default():
 
     # Run the function with default parameters
-    time_out, phase_out, density_out = density_through_orbit_theoretical()
+    time_out, phase_out, density_out = density_through_orbit_theoretical(resolution=0.01, show_plot=False, load_directly=True)
 
     # Assert outputs
     assert len(time_out) == len(phase_out) == len(density_out), "The output arrays should have the same length."
     assert isinstance(density_out, np.ndarray), "Density should be a NumPy array."
     assert len(density_out) > 0, "The density array should not be empty."
+    assert density_out.min() >= 0, "The density values should be non-negative."
 
-# Test for plot generation with default resolution
-def test_density_through_orbit_theoretical_plot(monkeypatch):
-    # Patch the internal functions
-    monkeypatch.setattr('xraybinaryorbit._manage_parameters', _mock_manage_parameters)
-    monkeypatch.setattr('xraybinaryorbit._orbital_phase_to_time', _mock_orbital_phase_to_time)
 
-    # Run the function with show_plot=True (it should not raise any exceptions)
-    try:
-        density_through_orbit_theoretical(show_plot=True)
-    except Exception as e:
-        pytest.fail(f"Plot generation failed with error: {e}")
-
-# Test for high resolution
-def test_density_through_orbit_theoretical_high_resolution(monkeypatch):
-    # Patch the internal functions
-    monkeypatch.setattr('xraybinaryorbit._manage_parameters', _mock_manage_parameters)
-    monkeypatch.setattr('xraybinaryorbit._orbital_phase_to_time', _mock_orbital_phase_to_time)
+# Test for higher resolution
+def test_density_through_orbit_theoretical_high_resolution():
 
     # Run the function with higher resolution
-    time_out, phase_out, density_out = density_through_orbit_theoretical(resolution=0.001)
+    time_out, phase_out, density_out = density_through_orbit_theoretical(resolution=0.001, show_plot=False, load_directly=True)
 
     # Assert outputs
     assert len(time_out) == len(phase_out) == len(density_out), "The output arrays should have the same length."
     assert isinstance(density_out, np.ndarray), "Density should be a NumPy array."
-    assert len(density_out) > 0, "The density array should not be empty."
-    assert len(phase_out) >= 1000, "The phase array should have more points with higher resolution."
[…]
