+++ test/test_timing_related_functions.py
@@ -1,5 +1,8 @@
 import pytest
 import numpy as np
+import sys
+
+sys.path.append('/Users/graci/Desktop/git/xraybinaryorbit/xraybinaryorbit')
 from xraybinaryorbit import *
 
 # Fake light curve generation
