+++ examples/example_fexxv_doppler_shifts_CenX3/bounds_orbit_bounds.txt
@@ -1,0 +1,3 @@
+iphase,semimajor,orbitalperiod,eccentricity,periapsis,inclination,Rstar,Mstar1,Mstar2,wind_vel,feature
+0.0,0.1,0.1,0.0,0.0,68.0,11.8,18.7,1.33,0.0,6.63
+1.0,3.0,2.0,1e-05,1e-05,76.0,12.2,22.77,1.45,1.0,6.7

[…]
