+++ xraybinaryorbit/xraybinaryorbit.py
@@ -562,68 +562,6 @@ def _manage_parameters(param_list, name, load_directly=False):
 
 #................................................... Easy way to manage bounds in fitting functions
 
-def _load_bounds_to_interface(param_list, lower_bounds, upper_bounds, name):
-    """
-    Create a graphical user interface (GUI) for inputting and updating lower and upper bounds for a list of parameters.
-
-    This function displays a Tkinter-based GUI where users can enter or modify both the lower and upper bounds for
-    a given list of parameters. It pre-populates the input fields with the previously provided `lower_bounds` and
-    `upper_bounds` values. Upon submission, the new bounds are saved and the GUI is closed.
-
-    Parameters
-    ----------
-    param_list : list of str
-        A list of parameter names for which the bounds are to be set.
-    lower_bounds : list of float
-        The initial lower bounds for the parameters, used to pre-fill the corresponding input fields.
-    upper_bounds : list of float
-        The initial upper bounds for the parameters, used to pre-fill the corresponding input fields.
-    name : str
-        The name for the form, used as the title of the Tkinter window.
-
-    Returns
-    -------
-    None
-        The function updates the global variables `lower_bounds_list` and `upper_bounds_list` with the new values
-        entered by the user.
-    """
-    
-    global lower_bounds_list, upper_bounds_list
-    
-    lower_bounds_list = lower_bounds
-    upper_bounds_list = upper_bounds
-
-    root = tk.Tk()
-    root.title("Bounds Input Form")
-
-    # Create entry fields for lower bounds
-    lower_entries = {}
-    for i, param in enumerate(param_list):
-        tk.Label(root, text=f"Lower {param}").grid(row=i, column=0)
-        lower_entries[param] = tk.Entry(root)
-        lower_entries[param].insert(0, str(lower_bounds_list[i]))  # Insert previous value
-        lower_entries[param].grid(row=i, column=1)
-
-    # Create entry fields for upper bounds
-    upper_entries = {}
-    for i, param in enumerate(param_list):
-        tk.Label(root, text=f"Upper {param}").grid(row=i, column=2)
-        upper_entries[param] = tk.Entry(root)
-        upper_entries[param].insert(0, str(upper_bounds_list[i]))  # Insert previous value
-        upper_entries[param].grid(row=i, column=3)
-
-    # Create submit button
-    def submit_form_bounds():
-        global lower_bounds_list, upper_bounds_list
-        lower_bounds_list = [float(lower_entries[param].get()) for param in param_list]
-        upper_bounds_list = [float(upper_entries[param].get()) for param in param_list]
-        root.withdraw()
-        root.destroy() # Close the root window after submitting the form
-
-    submit_button = tk.Button(root, text="Submit", command=submit_form_bounds)
-    submit_button.grid(row=len(param_list), columnspan=4)
-
-    root.mainloop()
     
 def _load_bounds_to_interface(param_list, lower_bounds, upper_bounds, name):
     """
@@ -644,7 +582,6 @@ def _load_bounds_to_interface(param_list, lower_bounds, upper_bounds, name):
         The name of the form or window, used as the title of the Tkinter window.
 
     """
-
 
     global lower_bounds_list, upper_bounds_list
     
@@ -2168,7 +2105,7 @@ def _conic_orbit(x_data, iphase, semimajor, orbitalperiod, eccentricity, periaps
[…]
