+++ test/test_fit_ps_and_ls.py
@@ -28,35 +28,35 @@ def test_fit_orbit_ps():
     assert len(predicted_data) == len(x), "Predicted data should match input length."
     assert isinstance(r_squared, float), "R-squared should be a float."
     
-def test_fit_spiral_in_orbit_ps():
+#def test_fit_spiral_in_orbit_ps():
     # Run the function
-    result, phase, predicted_data, r_squared = fit_spiral_in_orbit_ps(x, y, y_err=y_err, num_iterations=1, maxiter=100, swarmsize=10, units="keV", method_="extended", extended_binsize=0.01, load_directly=True)
+ #   result, phase, predicted_data, r_squared = fit_spiral_in_orbit_ps(x, y, y_err=y_err, num_iterations=1, maxiter=100, swarmsize=10, units="keV", method_="extended", extended_binsize=0.01, load_directly=True)
 
     # Assert the result is a pandas DataFrame
-    assert isinstance(result, pd.DataFrame), "Result should be a pandas DataFrame."
-    assert len(phase) == len(x), "Phase should match input length."
-    assert len(predicted_data) == len(x), "Predicted data should match input length."
-    assert isinstance(r_squared, float), "R-squared should be a float."
+   # assert isinstance(result, pd.DataFrame), "Result should be a pandas DataFrame."
+   # assert len(phase) == len(x), "Phase should match input length."
+    #assert len(predicted_data) == len(x), "Predicted data should match input length."
+    #assert isinstance(r_squared, float), "R-squared should be a float."
     
-def test_fit_spiral_ps():
+#def test_fit_spiral_ps():
     # Run the function
-    result, phase, predicted_data, r_squared = fit_spiral_ps(x, y, y_err=y_err, num_iterations=1, maxiter=100, swarmsize=10, units="keV", method_="extended", extended_binsize=0.01, load_directly=True)
+ #   result, phase, predicted_data, r_squared = fit_spiral_ps(x, y, y_err=y_err, num_iterations=1, maxiter=100, swarmsize=10, units="keV", method_="extended", extended_binsize=0.01, load_directly=True)
 
     # Assert the result is a pandas DataFrame
-    assert isinstance(result, pd.DataFrame), "Result should be a pandas DataFrame."
-    assert len(phase) == len(x), "Phase should match input length."
-    assert len(predicted_data) == len(x), "Predicted data should match input length."
-    assert isinstance(r_squared, float), "R-squared should be a float."
+   # assert isinstance(result, pd.DataFrame), "Result should be a pandas DataFrame."
+  #  assert len(phase) == len(x), "Phase should match input length."
+   # assert len(predicted_data) == len(x), "Predicted data should match input length."
+   # assert isinstance(r_squared, float), "R-squared should be a float."
     
-def test_fit_disc_ps():
+#def test_fit_disc_ps():
     # Run the function
-    result, phase,dphase, predicted_data, r_squared = fit_disc_ps(x, y, y_err=y_err, num_iterations=1, maxiter=100, swarmsize=10, units="keV", method_="extended", extended_binsize=0.01, load_directly=True)
+#    result, phase,dphase, predicted_data, r_squared = fit_disc_ps(x, y, y_err=y_err, num_iterations=1, maxiter=100, swarmsize=10, units="keV", method_="extended", extended_binsize=0.01, load_directly=True)
 
     # Assert the result is a pandas DataFrame
-    assert isinstance(result, pd.DataFrame), "Result should be a pandas DataFrame."
-    assert len(phase) == len(x), "Phase should match input length."
-    assert len(predicted_data) == len(x), "Predicted data should match input length."
-    assert isinstance(r_squared, float), "R-squared should be a float."
[…]
