+++ examples/example_fexxv_doppler_shifts_CenX3/bounds_spiral.txt
@@ -1,3 +1,3 @@
 iphase_spiral,semimajor_spiral,b,omega,inclination_spiral,feature
-nan,nan,nan,nan,nan,nan
-nan,nan,nan,nan,nan,nan
+0.0,1.0,-1.0,0.0,0.0,6.63
+1.0,100.0,0.0,1.0,90.0,6.7
