+++ examples/example_fexxv_doppler_shifts_CenX3/ionization_map_phase.txt
@@ -1,2 +1,0 @@
-phase,semimajor,eccentricity,periapsis,Rstar,Mstar1,Mstar2,wind_infinite_velocity,Mass_loss_rate,beta,luminosity,bound1,bound2
-0.1,3.0,0.6,200.0,12.0,22.0,1.4,1000.0,1e-06,0.8,1e-34,10.0,100.0
