+++ absorption_column_through_orbit.txt
@@ -1,0 +1,2 @@
+semimajor,orbitalperiod,eccentricity,periapsis,inclination,Rstar,Mstar1,Mstar2,wind_infinite_velocity,Mass_loss_rate,beta
+2.0,1.0,0.2,100.0,56.0,22.0,22.0,1.4,1200.0,1e-05,0.8
