+++ spiral.txt
@@ -1,3 +1,2 @@
 iphase_spiral,semimajor_spiral,b,omega,inclination_spiral,feature
-0.0,0.1,-0.1,1e-08,67.0,6.63
-1.0,20.0,0.0,0.01,77.0,6.7
+0.0,20.0,-0.01,0.0001,67.0,6.4
