+++ xraybinaryorbit/xraybinaryorbit.py
@@ -688,11 +688,13 @@ def _manage_bounds(param_list, name, load_directly=False):
 
     # Try to load previous bounds if available
     try:
-        with open(f"bounds_{name}.txt", "r") as file:
+        with open(f"{name}.txt", "r") as file:
             lines = file.readlines()
+
             param_names = lines[0].strip().split(",")  # Extract parameter names
             lower_bounds_list = [float(val) for val in lines[1].strip().split(",")]
             upper_bounds_list = [float(val) for val in lines[2].strip().split(",")]
+            
     except FileNotFoundError:
         file_exists = False
         param_names = param_list
@@ -704,7 +706,7 @@ def _manage_bounds(param_list, name, load_directly=False):
         _load_bounds_to_interface(param_names, lower_bounds_list, upper_bounds_list, name)
 
     # Save updated bounds to the file
-    with open(f"bounds_{name}.txt", "w") as file:
+    with open(f"{name}.txt", "w") as file:
         file.write(",".join(map(str, param_names)) + "\n")  # Write parameter names
         file.write(",".join(map(str, lower_bounds_list)) + "\n")  # Write lower bounds
         file.write(",".join(map(str, upper_bounds_list)) + "\n")  # Write upper bounds
@@ -2237,7 +2239,9 @@ def fit_orbit_ps(x_data, y_data, y_err=0, num_iterations=3, maxiter=1000, swarms
         return chi_squared
 
     #............................................PS implementation
-    lower_bounds, upper_bounds = _manage_bounds(parameter_names, "orbit",load_directly=load_directly)
+    lower_bounds, upper_bounds = _manage_bounds(parameter_names, "bounds_orbit", load_directly=load_directly)
+
+
     best_params_list = []
     chi_list = []
     
@@ -2370,7 +2374,7 @@ def fit_orbit_ls(x_data, y_data, y_err=0, units="keV", method_="extended", exten
         method_ = "discrete"
         
     #............................................LS implementation
-    lower_bounds, upper_bounds = _manage_bounds(parameter_names, "orbit",load_directly=load_directly)
+    lower_bounds, upper_bounds = _manage_bounds(parameter_names, "bounds_orbit",load_directly=load_directly)
     
     model_func = lambda x_data, iphase, semimajor, orbitalperiod, eccentricity, periapsis, inclination, Rstar, Mstar1, Mstar2,wind_vel, feature: _conic_orbit(x_data, iphase, semimajor, orbitalperiod, eccentricity, periapsis, inclination, Rstar, Mstar1, Mstar2, wind_vel,feature, units=units, method_=method_,extended_binsize=extended_binsize)
     
@@ -2701,7 +2705,7 @@ def fit_disc_ps(x_data, y_data, y_err=0, num_iterations=3, maxiter=1000, swarmsi
         return chi_squared
         
     #............................................PS implementation
-    lower_bounds, upper_bounds = _manage_bounds(parameter_names, "disc",load_directly=load_directly)
+    lower_bounds, upper_bounds = _manage_bounds(parameter_names, "bounds_disc",load_directly=load_directly)
     best_params_list = []
     chi_list = []
     
@@ -2855,7 +2859,7 @@ def fit_disc_ls(x_data, y_data, y_err=0, units="keV", method_="extended", extend
         method_ = "discrete"
         
     #............................................LS implementation
-    lower_bounds, upper_bounds = _manage_bounds(parameter_names, "disc",load_directly=load_directly)
[…]
