+++ test/test_fit_ps_and_ls.py
@@ -1,123 +1,93 @@
 import pytest
 import numpy as np
 import sys
+import pandas as pd
 
+# Adding the system path for module imports
 sys.path.append('/Users/graci/Desktop/git/xraybinaryorbit/xraybinaryorbit')
 from xraybinaryorbit import *
+from xraybinaryorbit import _manage_parameters, _orbital_phase_to_time, _orbital_time_to_phase, _manage_bounds
 
-
-x = np.array([[7.97000557e+08, 7.97002557e+08],
-       [7.97004557e+08, 7.97006557e+08]])
-
+# Test data
+x = np.array([[7.97000557e+08, 7.97002557e+08], [7.97004557e+08, 7.97006557e+08]])
 y = np.array([6.68379, 6.80991])
-
-y_err = np.array([[0.01962, 0.01599],[0.09961, 0.01009]])
-
-
-
-###############################PS
-
+y_err = np.array([[0.01962, 0.01599], [0.09961, 0.01009]])
 
 def test_fit_orbit_ps():
     # Run the function
-    result, phase, predicted_data, r_squared = fit_orbit_ps(x, y, y_err=y_err, num_iterations=1, maxiter=100, swarmsize=10, units="keV", method_="extended", extended_binsize=0.01, load_directly=True)
+    result, phase, predicted_data, r_squared = fit_orbit_ps(
+        x, y, y_err=y_err, num_iterations=1, maxiter=100,
+        swarmsize=10, units="keV", method_="extended",
+        extended_binsize=0.01, load_directly=True
+    )
 
-    # Assert the result is a pandas DataFrame
-    assert isinstance(result, pd.DataFrame), "Result should be a pandas DataFrame."
-    assert len(phase) == len(x), "Phase should match input length."
-    assert len(predicted_data) == len(x), "Predicted data should match input length."
-    assert isinstance(r_squared, float), "R-squared should be a float."
-    
-#def test_fit_spiral_in_orbit_ps():
-    # Run the function
- #   result, phase, predicted_data, r_squared = fit_spiral_in_orbit_ps(x, y, y_err=y_err, num_iterations=1, maxiter=100, swarmsize=10, units="keV", method_="extended", extended_binsize=0.01, load_directly=True)
-
-    # Assert the result is a pandas DataFrame
-   # assert isinstance(result, pd.DataFrame), "Result should be a pandas DataFrame."
-   # assert len(phase) == len(x), "Phase should match input length."
-    #assert len(predicted_data) == len(x), "Predicted data should match input length."
-    #assert isinstance(r_squared, float), "R-squared should be a float."
-    
-#def test_fit_spiral_ps():
-    # Run the function
- #   result, phase, predicted_data, r_squared = fit_spiral_ps(x, y, y_err=y_err, num_iterations=1, maxiter=100, swarmsize=10, units="keV", method_="extended", extended_binsize=0.01, load_directly=True)
-
-    # Assert the result is a pandas DataFrame
-   # assert isinstance(result, pd.DataFrame), "Result should be a pandas DataFrame."
-  #  assert len(phase) == len(x), "Phase should match input length."
-   # assert len(predicted_data) == len(x), "Predicted data should match input length."
-   # assert isinstance(r_squared, float), "R-squared should be a float."
-    
-#def test_fit_disc_ps():
-    # Run the function
-#    result, phase,dphase, predicted_data, r_squared = fit_disc_ps(x, y, y_err=y_err, num_iterations=1, maxiter=100, swarmsize=10, units="keV", method_="extended", extended_binsize=0.01, load_directly=True)
-
-    # Assert the result is a pandas DataFrame
- #   assert isinstance(result, pd.DataFrame), "Result should be a pandas DataFrame."
- #   assert len(phase) == len(x), "Phase should match input length."
[…]
