+++ examples/example_NS_doppler_shifts_CenX3/bounds_orbit_bounds.txt
@@ -1,3 +1,0 @@
-iphase,semimajor,orbitalperiod,eccentricity,periapsis,inclination,Rstar,Mstar1,Mstar2,wind_vel,feature
-0.0,1.54,2.087,0.0,0.0,67.0,11.7,18.8,1.2,0.0,4.7
-1.0,1.63,2.088,1e-06,1e-06,78.0,12.6,22.0,1.6,1e-10,4.9
