+++ density_through_orbit.txt
@@ -1,2 +1,0 @@
-semimajor,orbitalperiod,eccentricity,periapsis,Rstar,Mstar1,Mstar2,wind_infinite_velocity,Mass_loss_rate,beta
-1.7,2.5,0.5,200.0,22.0,21.0,1.4,1500.0,1e-06,0.8
