+++ disc.txt
@@ -1,2 +1,0 @@
-iphase,semimajor,orbitalperiod,eccentricity,periapsis,inclination,Rstar,Mstar1,Mstar2,iphase2,semimajor2,orbitalperiod2,eccentricity2,periapsis2,inclination2,Mass3,wind_vel,feature
-0.0,2.1,3.0,0.3,200.0,78.0,22.0,24.0,1.4,0.2,0.2,0.13,0.3,200.0,90.0,0.001,700.0,6.4
