+++ bounds_orbit.txt
@@ -1,3 +1,0 @@
-iphase,semimajor,orbitalperiod,eccentricity,periapsis,inclination,Rstar,Mstar1,Mstar2,wind_vel,feature
-nan,nan,nan,nan,nan,nan,nan,nan,nan,nan,nan
-nan,nan,nan,nan,nan,nan,nan,nan,nan,nan,nan
