+++ examples/example_NS_doppler_shifts_CenX3/orbit.txt
@@ -1,2 +1,0 @@
-iphase,semimajor,orbitalperiod,eccentricity,periapsis,inclination,Rstar,Mstar1,Mstar2,wind_vel,feature
-nan,nan,nan,nan,nan,nan,nan,nan,nan,nan,nan
