### 2024-10-07 d70d39a
Update test_doppler_disc_theoretical.py
