### 2024-10-07 f8aca6a
tests and examples
