+++ test/test_absorption_column_through_orbit_theoretical.py
@@ -1,9 +1,4 @@
 import pytest
-import numpy as np
-import os
-import sys
-
-sys.path.append('/Users/graci/Desktop/git/xraybinaryorbit/xraybinaryorbit')
 from xraybinaryorbit import *
 
 # Test for the default case with plot generation disabled
