+++ .github/workflows/ci.yml
@@ -34,7 +34,8 @@ jobs:
       run: pytest --cache-clear    
 
     - name: Run tests with Xvfb
-      run: cd test & pytest --verbose --junitxml=pytest_results.xml
+      run: ls -lrt
+      #run: cd test & pytest --verbose --junitxml=pytest_results.xml
 
     - name: Upload test results
       if: always()
