+++ .github/workflows/ci-test.yml
@@ -30,7 +30,7 @@ jobs:
     # run: pytest --rootdir=./test/ -cache-clear    
 
     - name: Run tests with Xvfb
-      run: pytest --rootdir=./test/ --verbose --junitxml=pytest_results.xml
+      run: cd test & pytest --verbose --junitxml=pytest_results.xml
 
     - name: Upload test results
       if: always()
