+++ .github/workflows/ci-test.yml
@@ -25,7 +25,8 @@ jobs:
         pip install --upgrade xraybinaryorbit
         pip install pytest>=6.0 #pytest-xvfb
         pip install -r requirements.txt
-
+    - name: ls
+      run: pwd ; ls ; exit 1
     #- name: Clear pytest cache
     # run: pytest --rootdir=./test/ -cache-clear    
 
