+++ .github/workflows/ci-test.yml
@@ -8,7 +8,7 @@ jobs:
 
     strategy:
       matrix:
-        python-version: [3.8, 3.9]
+        python-version: [3.8]
 
     steps:
     - name: Checkout code
@@ -26,7 +26,7 @@ jobs:
         pip install pytest>=6.0 #pytest-xvfb
         pip install -r requirements.txt
     - name: ls
-      run: pwd ; ls ; exit 1
+      run: pwd ; ls
     #- name: Clear pytest cache
     # run: pytest --rootdir=./test/ -cache-clear    
 
