+++ .github/workflows/ci-test.yml
@@ -31,7 +31,7 @@ jobs:
     # run: pytest --rootdir=./test/ -cache-clear    
 
     - name: Run tests with Xvfb
-      run: cd test & pytest --rootdir=test/ --verbose --junitxml=pytest_results.xml
+      run: cd test & pytest --verbose --junitxml=pytest_results.xml
 
     - name: Upload test results
       if: always()
