+++ xraybinaryorbit/xraybinaryorbit.py
@@ -1535,10 +1535,112 @@ def absorption_column_through_orbit_theoretical(resolution=0.01, show_plot=True,
 
         plt.tight_layout()
         
-        plt.savefig("NH_through_the_orbir.png")
+        plt.savefig("NH_through_the_orbit.png")
     #...................................................
     
     return time, th, nh
+         
+         
+
+# DENSITY AND LOGCHI #############################################################################
+
+def density_and_ionizationpar_in_ophase_theoretical(resolution=0.01, size=10, show_plot=True, load_directly=False):
+    """
+    Calculates and visualizes the density and ionization parameter (log(ξ)) encountered by radiation emitted
+    at each orbital phase as it travels towards an observer. Assumes a spherically distributed, neutral stellar
+    wind based on the CAK (Castor, Abbott, Klein) model. The density profile and ionization parameter are calculated
+    along the path from a neutron star (NS) through the stellar wind of its companion.
+
+    Parameters
+    ----------
+    resolution : float, optional
+        The resolution for the phase array (orbital phase). Default is 0.01.
+    size : float, optional
+        The scaling factor for the size of the path from the NS through the stellar wind. Determines how far into
+        the wind the path is extended. Default is 10.
+    show_plot : bool, optional
+        If True, displays and saves plots of the density, ionization parameter, and orbital path. Default is True.
+    load_directly : bool, optional
+        If True, attempts to load previously saved orbital parameters from a file. If False, prompts the user to
+        input the parameters. Default is False.
+
+    Returns
+    -------
+    z : array-like
+        Array of distances along the path from the neutron star to the observer (in cm).
+    density : array-like
+        Density profile of the stellar wind (in cm$^{-3}$) along the path.
+    chi : array-like
+        Ionization parameter (log(ξ)) calculated at each point along the path.
+
+    """
+
+    parameter_names = [ "orb_phase", "luminosity","semimajor", "eccentricity", "periapsis", "inclination","Rstar", "Mstar1", "Mstar2", "wind_infinite_velocity", "Mass_loss_rate", "beta"]
+    
+    fixed_values = _manage_parameters(parameter_names, "den_chi_orbphase",load_directly=load_directly)
+    orb_phase, luminosity, semimajor, eccentricity, periapsis, inclination ,Rstar, Mstar1, Mstar2, wind_infinite_velocity, Mass_loss_rate, beta = fixed_values
+
+    th = np.arange(0, 1, resolution)
+    abar = semimajor * max(Mstar1, Mstar2) / (Mstar1 + Mstar2)#BARICENTER CORRECTION
+    
+    luminosity_ = luminosity * 1e+32
+    M_dot_grams = Mass_loss_rate * msun / (365 * 24 * 60 * 60)  # Mass loss rate in grams/s
+    vinf_cm_s = wind_infinite_velocity * 100000  # Wind velocity in cm/s
+    Rstar_cm = Rstar * rsun_cm  # Convert stellar radius to cm
+
+    # Calculate orbital radius
+    Rorb_plot = (abar * (1 - eccentricity**2) / (1 + eccentricity * np.cos((th - periapsis / 360) * 2 * np.pi))) * Rstar * rsun_cm  # In cm
+    Rorb = (abar * (1 - eccentricity**2) / (1 + eccentricity * np.cos((orb_phase - periapsis / 360) * 2 * np.pi))) * Rstar * rsun_cm  # In cm
+    
[…]
