### 2024-10-15 #7220 @warrickball
Hi @matteobachetti & @tddesjardins, I was wondering if you'd made any progress with your reviews since I last checked in? I see there's been some activity in some issues but @xragua has responded in each case (and some can possibly even be closed).
