+++ .github/workflows/ci-test.yml
@@ -8,7 +8,7 @@ jobs:
 
     strategy:
       matrix:
-        python-version: [3.8]
+        python-version: [3.8, 3.10, 3.12]
 
     steps:
     - name: Checkout code
