+++ .github/workflows/ci-test.yml
@@ -2,6 +2,12 @@ name: Run CI-test
 
 on:
   push:
+    branches:
+    - main
+  pull_request:
+  schedule:
+    # run every Monday at 6am UTC
+    - cron: '0 6 * * 1'
 jobs:
   test:
     runs-on: ubuntu-latest
