### 2024-10-16 0105f46
Update README.md
Added a new function description
