+++ .github/workflows/ci-test.yml
@@ -14,7 +14,7 @@ jobs:
 
     strategy:
       matrix:
-        python-version: [3.8, 3.10, 3.12]
+        python-version: [3.8, 3.11, 3.12]
 
     steps:
     - name: Checkout code
