### 2024-10-16 e785101
Update ci-test.yml
Add strings  python-version: ["3.8", "3.10", "3.12"]
