+++ xraybinaryorbit/xraybinaryorbit.py
@@ -66,7 +66,8 @@ rsun_cm = 696340*1000*100 #cm
 kev_ams = 1.23984193
 
 na = 6.02214076*10**23/1.00797
-mp=0.5
+mu = 0.5
+mp = 1.67E-24
 
 #Helper functions
 ##########################################################################################
