### 2024-10-16 8943afb
unify inclination criteria
