### 2024-10-16 #7220 @xragua
Dear @matteobachetti and @tddesjardins, I just wanted to mention that I have added a small function that fits well this package (density_and_ionization_orbital_phase_theoretical), its purpose is to calculate the density and ionization parameter the radiation emitted by the NS finds from the point that is emitted in the orbit in its path to the observer until certain extent (default 10 R*). It has been added in the documentation, in the package of course, added to the CI tests and in the examples section. Thank you!
