+++ index.rst
@@ -1,0 +1,20 @@
+.. xraybinaryorbit documentation master file, created by
+   sphinx-quickstart on Wed Oct 16 16:03:06 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to xraybinaryorbit's documentation!
+===========================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
