### 2024-10-17 fb24537
Delete index.rst
