+++ docs/_sources/installation.rst.txt
@@ -1,58 +1,0 @@
-Installation
-============
-
-This section describes how to install the **xraybinaryorbit** package.
-
-
- Installing via pip
- ~~~~~~~~~~~~~~~~~~~~~
-
-To install **xraybinaryorbit** from the Python Package Index (PyPI), you can use `pip`, the Python package installer. Run the following command in your terminal or command prompt:
-
-.. code-block:: bash
-
-   pip install xraybinaryorbit
-
-This will automatically install the latest stable version of **xraybinaryorbit** along with its dependencies.
-
- Installing from Source
- ~~~~~~~~~~~~~~~~~~~~~
-
-If you prefer to install the package from source, you can clone the repository from GitHub and install it manually:
-
-1. Clone the repository:
-
-   .. code-block:: bash
-
-      git clone https://github.com/xragua/xraybinaryorbit
-      
-2. Navigate to the package directory:
-
-   .. code-block:: bash
-
-      cd xraybinaryorbit
-
-3. Install the package using `pip`:
-
-   .. code-block:: bash
-
-      pip install .
-
- Dependencies
- ~~~~~~~~~~~~~~~~~~~~~
-
-**xraybinaryorbit** requires the following Python libraries:
-
-- `numpy`
-- `pandas`
-- `scipy`
-- `pyswarm`
-- `matplotlib`
-- `os`
-- `warnings`
-- `tkinter`
-- `astropy`
-- `inspect`
-
-These dependencies are automatically installed when you install the package via `pip`. If you encounter any issues, ensure that these libraries are installed on your system.
-
