+++ docs/usage.rst
@@ -1,617 +1,0 @@
-.. xraybinaryorbit documentation master file, created by
-   sphinx-quickstart on Mon Sep  9 18:12:16 2024.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Usage
-=====
-
-
-Theoretical Functions
-~~~~~~~~~~~~~~~~~~~~~
-
-Below, the followinf functions are described:
-
-- **doppler_orbit_theoretical**: Calculates the Doppler effect in the orbital motion.
-- **doppler_spiral_theoretical**: Models the Doppler effect in spiral structures.
-- **doppler_disc_theoretical**: Models the Doppler effect in accretion discs. 
-- **doppler_spiral_in_orbit_theoretical**: Combines orbital and spiral Doppler effects. 
-- **density_through_orbit_theoretical**: Computes wind density through the orbit. 
-- **absorption_column_through_orbit_theoretical**: Calculates absorption column variations through the orbit. 
-- **ionization_map_phase**: Maps ionization levels across orbital phases. 
-- **orbital_phase_to_time**: Converts orbital phase to time.
-- **orbital_time_to_phase**: Converts orbital time to phase. 
-
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-doppler_orbit_theoretical
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-    Computes the Doppler variation expected from orbital movement given a time array in seconds.
-
-    Parameters:
-    - t (array-like): Time array in seconds.
-    - units (str, optional): Units for the output. Default is "keV". Can be "keV", "s", or "angstrom".
-    - show_plot (bool, optional): If True, a plot of the orbit and Doppler evolution is shown and saved. Default is False.
-    - precision_for_phase (float, optional): Precision for phase calculation. Default is 0.01.
-
-    Returns:
-    - t (array-like): Input time array.
-    - x (array-like): Orbital phase array.
-    - equation (array-like): Expected Doppler variation.
-    
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-doppler_spiral_theoretical
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-    Computes the Doppler variation expected from a spiral movement given a time array in seconds.
-    
-    A logarithmic spiral is a type of spiral that grows in size by a constant factor with each turn. Its equation in polar coordinates is
-    r = a * e^(b * θ), where:
-
-    r is the distance from the origin (radius)
-    omega is the angle from a reference direction (usually the positive x-axis)
-    a is the scale factor that determines how quickly the spiral grows
-    b is the rate of rotation, controlling the tightness or looseness of the spiral
-
-    Computes the Doppler variation expected from a spiral movement given a time array in seconds.
-
-    Parameters:
-    - t (array-like): Time array in seconds.
-    - units (str, optional): Units for the output. Default is "keV". Can be "keV", "s", or "angstrom".
-    - show_plot (bool, optional): If True, a plot of the spiral and Doppler evolution is shown and saved. Default is False.
-    - A form will appear to input the necessary orbital parameters. These parameters will be saved in a .txt file in the current directory and automatically loaded in subsequent runs. This avoids the need to re-enter all parameters; you can modify only those that require adjustment.
-    
-    Returns:
-    - x (array-like): Orbital phase array.
-    - equation (array-like): Expected Doppler variation.
-    
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-doppler_disc_theoretical
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-    Computes the Doppler variation expected from an orbital movement in a main orbit,
[…]
