+++ my-project/docs/index.md
@@ -1,0 +1,611 @@
+# xraybinaryorbit
+
+**xraybinaryorbit** is a user-friendly Python package designed to facilitate the analysis of orbital modulations in X-ray binaries.
+
+Although orbital modulations are well-documented, their analysis is complex and dependent on various parameters and geometrical considerations. This package is designed to be useful for nearly every X-ray binary analysis, aiming to facilitate its implementation for other astronomers.
+
+Looking ahead, **high-resolution x-ray spectroscopy** will significantly rely on these analyses to provide deep insights into the orbital dynamics of X-ray binary systems.
+
+Here, we propose a user-friendly "form" method to improve the package usability. These forms, once loaded and submited, are saved for future interactions.
+
+If you need something or have any doubts, please, write me: graciela.sanjurjo@ua.es
+
+### Installation
+
+You can install the package directly from PyPI using `pip`:
+pip install xraybinaryorbit 
+or dowload the code from (https://github.com/xragua/xraybinaryorbit/releases/tag/0.2.9)
+In (https://github.com/xragua/xraybinaryorbit/tree/main/examples) some examples of their usage are presented.
+
+
+## FUNCTIONS:
+
+###Theoretical Functions:
+
+These functions will enable us to model the theoretical behavior of various astrophysical features influenced by orbital modulations with the aim of comparing our observational results with the expected.
+
+####doppler_orbit_theoretical
+-------------------------------------------------------------------------------
+    Computes the Doppler variation expected from orbital movement given a time array in seconds.
+
+    Parameters:
+    - t (array-like): Time array in seconds.
+    - units (str, optional): Units for the output. Default is "keV". Can be "keV", "s", or "angstrom".
+    - show_plot (bool, optional): If True, a plot of the orbit and Doppler evolution is shown and saved. Default is False.
+    - precision_for_phase (float, optional): Precision for phase calculation. Default is 0.01.
+
+    Returns:
+    - t (array-like): Input time array.
+    - x (array-like): Orbital phase array.
+    - equation (array-like): Expected Doppler variation.
+----------------------------------------------------------------------------------
+####doppler_spiral_theoretical
+----------------------------------------------------------------------------------
+    Computes the Doppler variation expected from a spiral movement given a time array in seconds.
+    
+    A logarithmic spiral is a type of spiral that grows in size by a constant factor with each turn. Its equation in polar coordinates is
+    r = a * e^(b * θ), where:
+
+    r is the distance from the origin (radius)
+    omega is the angle from a reference direction (usually the positive x-axis)
+    a is the scale factor that determines how quickly the spiral grows
+    b is the rate of rotation, controlling the tightness or looseness of the spiral
+
+    Computes the Doppler variation expected from a spiral movement given a time array in seconds.
+
+    Parameters:
+    - t (array-like): Time array in seconds.
+    - units (str, optional): Units for the output. Default is "keV". Can be "keV", "s", or "angstrom".
+    - show_plot (bool, optional): If True, a plot of the spiral and Doppler evolution is shown and saved. Default is False.
+    - A form will appear to input the necessary orbital parameters. These parameters will be saved in a .txt file in the current directory and automatically loaded in subsequent runs. This avoids the need to re-enter all parameters; you can modify only those that require adjustment.
+    
+    Returns:
+    - x (array-like): Orbital phase array.
+    - equation (array-like): Expected Doppler variation.
+----------------------------------------------------------------------------------
+####doppler_disc_theoretical
+----------------------------------------------------------------------------------
+    Computes the Doppler variation expected from an orbital movement in a main orbit,
+    assuming a ballistic movement of plasma around a compact object or the movement of a mass
[…]
