+++ my-project/mkdocs.yml
@@ -1,0 +1,3 @@
+site_name: XRAYBINARYORBIT
+theme: readthedocs
+highlights: true
