### 2024-10-17 0c5f372
Update static.yml
