+++ my-project/docs/index.md
@@ -1,5 +1,4 @@
 # xraybinaryorbit
-LAEX 
 
 **xraybinaryorbit** is a user-friendly Python package designed to facilitate the analysis of orbital modulations in X-ray binaries.
 
@@ -15,8 +14,8 @@ If you need something or have any doubts, please, write me: graciela.sanjurjo@ua
 
 You can install the package directly from PyPI using `pip`:
 pip install xraybinaryorbit 
-or dowload the code from https://github.com/xragua/xraybinaryorbit/releases/tag/0.2.9
-In https://github.com/xragua/xraybinaryorbit/tree/main/examples some examples of their usage are presented.
+Or download the code from [here](https://github.com/xragua/xraybinaryorbit/releases/tag/0.2.9).
+Some examples of their usage are presented [here](https://github.com/xragua/xraybinaryorbit/tree/main/examples).
 
 
 ## FUNCTIONS:
@@ -609,4 +608,29 @@ In this section we collect some usefull timing functions. From them we can obtai
 ----------------------------------------------------------------------------------
 
 ## REFERENCES:
+
+1. Bonyadi, M. R., & Michalewicz, Z. (2017). Particle Swarm Optimization for Single Objective Continuous Space Problems: A Review. *Evolutionary Computation, 25*(1), 1-54. [https://doi.org/10.1162/EVCO_r_00180](https://doi.org/10.1162/EVCO_r_00180)
+
+2. Sanjurjo-Ferrín, G., Torrejón, J. M., Postnov, K., Oskinova, L., Rodes-Roca, J. J., & Bernabeu, G. (2021). X-ray variability of the HMXB Cen X-3: evidence for inhomogeneous accretion flows. *Monthly Notices of the Royal Astronomical Society, 501*(4), 5892-5909. [https://doi.org/10.1093/mnras/staa3953](https://doi.org/10.1093/mnras/staa3953)
+
+3. Barret, D., Lam Trong, T., den Herder, J.-W., & Piro. (2016). The Athena X-ray Integral Field Unit (X-IFU). In J.-W. A. den Herder, T. Takahashi, & M. Bautz (Eds.), *Space Telescopes and Instrumentation 2016: Ultraviolet to Gamma Ray* (Vol. 9905, p. 99052F). Society of Photo-Optical Instrumentation Engineers (SPIE). [https://doi.org/10.1117/12.2232432](https://doi.org/10.1117/12.2232432)
+
+4. Virtanen, P., Gommers, R., Oliphant, T. E., Haberland, M., Reddy, T., Cournapeau, D., Burovski, E., Peterson, P., Weckesser, W., Bright, J., van der Walt, S. J., Brett, M., Wilson, J., Jarrod Millman, K., Mayorov, N., Nelson, A. R. J., Jones, E., Kern, R., Larson, E., Carey, C. J., Polat, İ., Feng, Y., Moore, E. W., VanderPlas, J., Laxalde, D., Perktold, J., Cimrman, R., Henriksen, I., Quintero, E. A., Harris, C. R., Archibald, A. M., Ribeiro, A. H., Pedregosa, F., van Mulbregt, P., & Contributors, S. (2020). SciPy 1.0: Fundamental Algorithms for Scientific Computing in Python. *Nature Methods, 17*, 261-272. [https://doi.org/10.1038/s41592-019-0686-2](https://doi.org/10.1038/s41592-019-0686-2)
+
+5. PySwarms: A Python-Based Swarm Optimization Library. (2022). Retrieved from [https://pythonhosted.org/pyswarm/](https://pythonhosted.org/pyswarm/)
+
[…]
