### 2024-10-17 #7220 @matteobachetti
@xragua thanks. As I mentioned in my review, I think the docs should do some more than a list of functions with their API documentation. The documentation of a package of this kind, in my opinion, should introduce curious users to the kind of science analysis they can do with it, with a robust introduction (sort of an expanded statement of need!), and then maybe go through a couple of simple examples, the typical "getting started". The API documentation should be for those who are already using the code and want to understand the subtleties of it, at the end.
You have good notebooks, so I think it shouldn't be so difficult linking them in the documentation through `nbsphinx` and form a large section of examples, maybe ordered from the simplest to the most advanced
