+++ my-project/docs/functions.md
@@ -1,0 +1,590 @@
+
+## FUNCTIONS
+
+###Theoretical Functions
+
+These functions will enable us to model the theoretical behavior of various astrophysical features influenced by orbital modulations with the aim of comparing our observational results with the expected.
+
+####doppler_orbit_theoretical
+-------------------------------------------------------------------------------
+    Computes the Doppler variation expected from orbital movement given a time array in seconds.
+
+    Parameters:
+    - t (array-like): Time array in seconds.
+    - units (str, optional): Units for the output. Default is "keV". Can be "keV", "s", or "angstrom".
+    - show_plot (bool, optional): If True, a plot of the orbit and Doppler evolution is shown and saved. Default is False.
+    - precision_for_phase (float, optional): Precision for phase calculation. Default is 0.01.
+
+    Returns:
+    - t (array-like): Input time array.
+    - x (array-like): Orbital phase array.
+    - equation (array-like): Expected Doppler variation.
+----------------------------------------------------------------------------------
+####doppler_spiral_theoretical
+----------------------------------------------------------------------------------
+    Computes the Doppler variation expected from a spiral movement given a time array in seconds.
+    
+    A logarithmic spiral is a type of spiral that grows in size by a constant factor with each turn. Its equation in polar coordinates is
+    r = a * e^(b * θ), where:
+
+    r is the distance from the origin (radius)
+    omega is the angle from a reference direction (usually the positive x-axis)
+    a is the scale factor that determines how quickly the spiral grows
+    b is the rate of rotation, controlling the tightness or looseness of the spiral
+
+    Computes the Doppler variation expected from a spiral movement given a time array in seconds.
+
+    Parameters:
+    - t (array-like): Time array in seconds.
+    - units (str, optional): Units for the output. Default is "keV". Can be "keV", "s", or "angstrom".
+    - show_plot (bool, optional): If True, a plot of the spiral and Doppler evolution is shown and saved. Default is False.
+    - A form will appear to input the necessary orbital parameters. These parameters will be saved in a .txt file in the current directory and automatically loaded in subsequent runs. This avoids the need to re-enter all parameters; you can modify only those that require adjustment.
+    
+    Returns:
+    - x (array-like): Orbital phase array.
+    - equation (array-like): Expected Doppler variation.
+----------------------------------------------------------------------------------
+####doppler_disc_theoretical
+----------------------------------------------------------------------------------
+    Computes the Doppler variation expected from an orbital movement in a main orbit,
+    assuming a ballistic movement of plasma around a compact object or the movement of a mass
+    entering an accretion disc.
+
+    Parameters:
+    - t (array-like): Time array in seconds.
+    - units (str, optional): Units for the output. Default is "keV". Can be "keV", "s", or "angstrom".
+    - show_plot (bool, optional): If True, a plot of the disc and Doppler evolution is shown and saved. Default is False.
+    - A form will appear to input the necessary orbital parameters. These parameters will be saved in a .txt file in the current directory and automatically loaded in subsequent runs. This avoids the need to re-enter all parameters; you can modify only those that require adjustment.
+
+    Returns:
+    - t (array-like): Time array.
+    - x (array-like): Orbital phase array for the first orbit.
+    - x2 (array-like): Orbital phase array for the second orbit.
+    - equation (array-like): Expected Doppler variation.
+----------------------------------------------------------------------------------
+####doppler_spiral_in_orbit_theoretical
+----------------------------------------------------------------------------------
+    This function requires a time array in seconds and returns the time, orbital phase, and Doppler variation
+    expected under the assumption of an orbital movement with a logarithmic spiral component.
+
+    Parameters:
+    - t (array-like): Time array in seconds.
+    - units (str, optional): Units for the output. Default is "keV". Can be "keV", "s", or "angstrom".
[…]
