+++ my-project/docs/index.md
@@ -1,659 +1,44 @@
 # xraybinaryorbit
 
-**xraybinaryorbit** is a user-friendly Python package designed to facilitate the analysis of orbital modulations in X-ray binaries.
 
-Although orbital modulations are well-documented, their analysis is complex and dependent on various parameters and geometrical considerations. This package is designed to be useful for nearly every X-ray binary analysis, aiming to facilitate its implementation for other astronomers.
+**X-ray binaries are truly fascinating!** In these extreme environments, a compact object—either a neutron star or black hole—draws in matter from a companion star, producing intense X-ray emissions. These systems offer a unique window into extreme physics, from the effects of strong gravity and relativistic jets to the presence of intense magnetic fields.
 
-Looking ahead, **high-resolution x-ray spectroscopy** will significantly rely on these analyses to provide deep insights into the orbital dynamics of X-ray binary systems.
+Orbital modulations are observed in nearly all X-ray binary systems. These variations arise from the orbital motions of the system, driven by the relative velocities of the two stars and their changing configurations with respect to each other and the observer.
 
-Here, we propose a user-friendly "form" method to improve the package usability. These forms, once loaded and submited, are saved for future interactions.
+To aid in the study of these modulations, we introduce **xraybinaryorbit** —a user-friendly Python package designed to simplify the analysis of orbital modulations in X-ray binaries. Whether you are studying the orbital parameters or working to refine your data analysis, this tool is built to help you extract the most from your observations.
 
-If you need something or have any doubts, please, write me: graciela.sanjurjo@ua.es
+If you have any questions or need assistance, please feel free to reach out: graciela.sanjurjo@ua.es.
+
+
+
+## Getting started
 
 ### Installation
+---------------------------------------------------------
 
-You can install the package directly from PyPI using pip: **pip install xraybinaryorbit **
+
+You can install the package directly from PyPI using pip: **pip install xraybinaryorbit**.
 
 Or download the code from [here](https://github.com/xragua/xraybinaryorbit/releases/tag/0.2.9).
 
 Some examples of their usage are presented [here](https://github.com/xragua/xraybinaryorbit/tree/main/examples).
 
 
-## FUNCTIONS
+### Which orbital modulations are we talking about?
+---------------------------------------------------------
 
-###Theoretical Functions
+We can observe how the center of an emission line slightly changes during a phase resolved analysis, or how the NS spin period slightly varies following a trend. These phenomena can be caused by Doppler effect. Our code will help you turn these observations into the orbital parameters that cause those Doppler shifts.
 
-These functions will enable us to model the theoretical behavior of various astrophysical features influenced by orbital modulations with the aim of comparing our observational results with the expected.
+But this simple idea can get tricky when you consider all the factors involved. Inclination, eccentricity, periapsis, distance to the barycenter (which depends on the masses of the stars involved), and orbital phases really matter in this analysis. Plus, if there’s eccentricity different than 0, the velocity around the orbit isn’t constant. So, it’s not as straightforward as it might seem, right?.
 
-####doppler_orbit_theoretical
--------------------------------------------------------------------------------
-    Computes the Doppler variation expected from orbital movement given a time array in seconds.
+If we think about stellar wind (the matter accreted by our compact objects) there are many combinations that can result into orbital modulations. With the eccentricity the density around the orbit changes, and thus, the accreted matter. With eccentricity and inclination the absorption column faced by the emitted radiation varies depending on the orbital phase, so does the ionization of the wind.
 
-    Parameters:
-    - t (array-like): Time array in seconds.
-    - units (str, optional): Units for the output. Default is "keV". Can be "keV", "s", or "angstrom".
-    - show_plot (bool, optional): If True, a plot of the orbit and Doppler evolution is shown and saved. Default is False.
[…]

[…]
