+++ notebooks/example_fexxv_doppler_shifts_CenX3/bound_spiral_orbit.txt
@@ -1,0 +1,3 @@
+iphase_orbit,semimajor_orbit,orbitalperiod,eccentricity,periapsis,inclination,Rstar,Mstar1,Mstar2,iphase_spiral,semimajor_spiral,b,omega,inclination_spiral,feature
+nan,nan,nan,nan,nan,nan,nan,nan,nan,nan,nan,nan,nan,nan,nan
+nan,nan,nan,nan,nan,nan,nan,nan,nan,nan,nan,nan,nan,nan,nan
