+++ README.md
@@ -1,646 +1,85 @@
-# xraybinaryorbit
-
-xraybinaryorbit is a user-friendly Python package designed to facilitate the analysis of orbital modulations in X-ray binaries. 
-
-Although orbital modulations are widely known, they are complex to analyze and depend on several parameters and geometrical considerations. With this in mind, we collected all the functions we created through years of analyzing close X-ray binaries and formed a python package useful in almost every X-ray binary analysis, with the aim of facilitating its implementation for other astronomers. With the fact that these orbital modulations rely in several different parameters, here, we propose a user-friendly "form" method to improve the package usability. These forms, once loaded and submited, are saved for future interactions.
-
-Each function is briefly described when loaded, helping to guide its execution.
-
-The primary physics used for these packages rely on the Doppler effect, Kepler's laws, and the CAK model (Castor, Abbott, and Klein) as described in:
-Castor, J. I., Abbott, D. C., & Klein, R. I. (1975). Radiation-driven winds in Of stars. *Astrophysical Journal*, 195, 157-174.
-
-This package requires the following Python libraries to function correctly: numpy, pandas, scipy, pyswarm, matplotlib, os, warnings, tkinter, astropy, and inspect. These dependencies cover numerical computations, data manipulation, scientific computing, optimization, visualization, GUI creation, time-series analysis, and interaction with the operating system. Ensure these packages are installed to run this package successfully.
-
-### Installation
-
-You can install the package directly from PyPI using `pip`:
-pip install xraybinaryorbit
-
-### Available Functions
-
-#### Theoretical Functions
-- **doppler_orbit_theoretical**: Calculates the Doppler effect in the orbital motion.
-- **doppler_spiral_theoretical**: Models the Doppler effect in spiral structures.
-- **doppler_disc_theoretical**: Models the Doppler effect in accretion discs.
-- **doppler_spiral_in_orbit_theoretical**: Combines orbital and spiral Doppler effects.
-- **density_through_orbit_theoretical**: Computes wind density through the orbit.
-- **absorption_column_through_orbit_theoretical**: Calculates absorption column variations through the orbit.
-- **ionization_map_phase**: Maps ionization levels across orbital phases.
-- **orbital_phase_to_time**: Converts orbital phase to time.
-- **orbital_time_to_phase**: Converts orbital time to phase.
-
-#### Fitting Functions
-(Note: 'ps' stands for particle swarm optimization and 'ls' stands for least squares. Particle swarm optimization is preferred as least squares does not always converge.)
-- **fit_orbit_ps**: Fits orbital parameters using particle swarm optimization.
-- **fit_orbit_ls**: Fits orbital parameters using least squares.
-- **fit_disc_ps**: Fits disc parameters using particle swarm optimization.
-- **fit_disc_ls**: Fits disc parameters using least squares.
-- **fit_spiral_ps**: Fits spiral parameters using particle swarm optimization.
-- **fit_spiral_ls**: Fits spiral parameters using least squares.
-- **fit_spiral_in_orbit_ps**: Fits combined spiral and orbit parameters using particle swarm optimization.
-- **fit_spiral_in_orbit_ls**: Fits combined spiral and orbit parameters using least squares.
-- **nh_orbit**: Calculates the column density (N_H) variations through the orbit.
-- **fit_nh_ps**: Fits N_H variations using particle swarm optimization.
-
-#### Timing Functions
-- **hr**: Calculates hardness ratios.
-- **cr**: Computes count rates.
-- **rebin_snr**: Rebins data to achieve a specific signal-to-noise ratio.
-- **rebin_bins**: Rebins data into a specified number of bins.
-- **fold_pulse**: Folds pulse profiles over the orbital period.
[…]
