+++ contributing.md
@@ -1,0 +1,89 @@
+
+# Contributing to XRAYBINARYORBIT
+
+First off, thank you for taking the time to contribute! 🎉 We appreciate your effort in helping to improve the XRAYBINARYORBIT package.
+
+## How to Contribute
+
+### 1. Reporting Bugs
+
+If you find a bug, please open an issue [here](https://github.com/username/XRAYBINARYORBIT/issues). When reporting a bug, please include:
+- A clear and descriptive title.
+- A detailed description of the issue, including steps to reproduce it.
+- Any relevant error messages or stack traces.
+- Information about your system (OS, Python version, dependencies).
+- If possible, include code snippets that demonstrate the issue.
+
+### 2. Suggesting Enhancements
+
+Have a feature request or improvement idea? Please open a feature request [here](https://github.com/username/XRAYBINARYORBIT/issues) with:
+- A detailed description of the feature/enhancement.
+- Why it would be helpful to the project.
+- Any alternatives or considerations to keep in mind.
+
+### 3. Submitting Code Changes
+
+#### Step 1: Fork the Repository
+
+First, fork the repository to your own GitHub account. You can do this by clicking the "Fork" button on the repository’s main page.
+
+#### Step 2: Clone the Repository
+
+Once you’ve forked the repository, clone it to your local machine:
+
+```bash
+git clone https://github.com/your-username/XRAYBINARYORBIT.git
+cd XRAYBINARYORBIT
+```
+
+#### Step 3: Create a New Branch
+
+It’s a good idea to create a new branch for your changes:
+
+```bash
+git checkout -b feature/my-new-feature
+```
+
+#### Step 4: Make Your Changes
+
+Now you can start making your changes! Try to:
+- Follow the existing code style.
+- Add comments and documentation where appropriate.
+- Ensure your code is well-tested (see below).
+
+#### Step 5: Test Your Changes
+
+Before submitting your changes, please ensure they pass the existing tests and add any new tests for your feature/bugfix. You can run the tests using:
+
+```bash
+pytest tests/
+```
+
+#### Step 6: Commit Your Changes
+
+Once you’re happy with your changes, commit them with a meaningful message:
+
+```bash
+git add .
+git commit -m "Add feature XYZ to handle ABC"
+```
+
+#### Step 7: Push Your Changes
+
+Push your changes to your forked repository:
+
+```bash
+git push origin feature/my-new-feature
+```
+
+#### Step 8: Open a Pull Request
+
+Finally, open a pull request to the main repository. When submitting a pull request, please ensure:
+- The pull request has a clear title and description.
+- You’ve linked the pull request to any relevant issues.
+- Your code passes all tests.
+
+
+---
+
+Thank you for your contributions! 🙌
