+++ paper/paper.md
@@ -44,70 +44,84 @@ The xraybinaryoorbit package helps to unveil X-Ray binary orbital dynamics based
 
 # Science behind
 
-The functions contained in this package rely in the following theories:
+The package functions are based on the following key theories:
 
-## Conservation of angular momentum in orbital mechanics:
+### Conservation of Angular Momentum:
 
-If the eccentricity of our system is different than 0, the orbital phase will not vary linearly with the observational time, as the speed will increase at periastron primarily due to the conservation of angular momentum, which dictates that as the compact object moves closer to the central star, it must travel faster to maintain the total angular momentum of the system. This relationship is further influenced by Kepler’s laws of planetary motion, which describe how objects sweep out equal areas in equal times (see [@2006ima..book.....C] as an example).
+In systems with non-zero eccentricity, orbital phase varies non-linearly with observational time due to increased speed at periastron, governed by the conservation of angular momentum. Kepler’s laws further explain that objects sweep out equal areas in equal times (see [@2006ima..book.....C]).
 
 $$ r^2 \cdot \omega = h $$
 
-We will take this fact into consideration in all our functions and provide dedicated functions to transform phase into time and vice versa.
+We account for this in all functions and provide tools to convert phase to time and vice versa.
 
-## CAK model:
+### CAK Model:
 
-The CAK model, proposed by Castor, Abbott, and Klein in 1975 [@1975ApJ195157C], is a theoretical framework used to describe radiation-driven winds in massive stars. These stars have strong stellar winds driven by the interaction between their radiation and the surrounding material.
-
-The CAK model provides a quantitative description of how the wind velocity, density, and ionization state vary with distance from the companion.
+The CAK model [@1975ApJ195157C] describes radiation-driven winds in massive stars, with wind velocity, density, and ionization state varying with distance from the companion star.
 
 $$ \rho = \frac{\dot{M}}{4 \pi v R^2} $$
 
-where $\rho$ is the density of the wind at a given distance R $\dot{M}$ is the mass accretion rate in units of mass per unit of time, v is the orbital speed at distances greater than the stellar radius and R is the distance to the star.
-In this package, we assume that the wind is spherically distributed and unionized.
+Here, \(\rho\) is wind density, \(\dot{M}\) is the mass accretion rate, \(v\) is orbital speed, and \(R\) is distance to the star. We assume a spherical, unionized wind distribution.
 
-## Accretion Luminosity and Ionization Parameter:
+### Accretion Luminosity and Ionization:
 
-Accretion is the process by which gravitational potential energy is extracted from material accreting onto a gravitating body  (see [@Frank_King_Raine_2002]). This phenomenon serves as the primary power source in various types of close binary systems and is also believed to fuel active galactic nuclei and quasars. When considering a flux of matter with an accretion rate $\dot{M}$, the resulting luminosity (assuming all mechanical energy is radiated) is defined as the accretion luminosity:
+Accretion powers many close binary systems. The accretion luminosity is:
 
 $$ L_{ac} = \frac{GM \dot{M}}{R} $$
 
-where $L_{ac}$ is the accretion luminosity, G is the gravitational constant, M is the mass of the gravitating body, $\dot{M}$ is the accretion rate, and R is the characteristic radius associated with the accretion process.
-
-The ionization parameter $\xi$ is defined as:
+where \(L_{ac}\) is the luminosity, \(G\) is the gravitational constant, \(M\) is mass, and \(R\) is the accretion radius. The ionization parameter \(\xi\) is:
 
 $$ \xi = \frac{L_{\rm X}}{n(r_{\rm X}) r_{\rm X}^{2}} $$
 
[…]
