+++ my-project/mkdocs.yml
@@ -1,4 +1,4 @@
-site_name: GETTING STARTED
+site_name: xraybinaryorbit
 
 theme:
     palette:
@@ -13,28 +13,32 @@ extra_css:
 
 plugins:
     - mkdocs-jupyter
-    
-    
 
 nav:
-    - XRAYBINARYORBIT: index.md
+    -  GUETTING STARTED: index.md
     
-    - Notebooks:
-        - NS Spin Shape Vela X-1: notebooks/example_NS_spin_shape_VelaX1/Pulse_shape_Velax1.ipynb
+    - EXAMPLES:
+    
+
+        - Time Phase Relation Theoretical: notebooks/example_time_phase_relation_theoretical/time_phase_relation.ipynb
+    
+        - Doppler Theoretical in four scenarios: notebooks/example_doppler_theoretical/Doppler_theoretical_functions.ipynb
         
         - Density throuth the orbit line and NH Theoretical: notebooks/example_density_NH_theoretical/Wind_density_and_NH.ipynb
         
         - Density and ionization through the path of the emitted radiation depending on the orbital Phase: notebooks/example_density_ionization_orbital_phase/density_and_chi_orbital_phase.ipynb
         
-        - Doppler Theoretical in four scenarios: notebooks/example_doppler_theoretical/Doppler_theoretical_functions.ipynb
-        
-        - Fe XXV Doppler Shifts Cen X-3: notebooks/example_fexxv_doppler_shifts_CenX3/Fit_real_line_doppler_shifts.ipynb
-        
         - Ionization Map of 4U0114+65: notebooks/example_ionization_map_4U0114+65/Ionization_map.ipynb
         
-        - Time Phase Relation Theoretical: notebooks/example_time_phase_relation_theoretical/time_phase_relation.ipynb
+        - NS Spin Shape Vela X-1: notebooks/example_NS_spin_shape_VelaX1/Pulse_shape_Velax1.ipynb
+
+        - Fe XXV Doppler Shifts Cen X-3: notebooks/example_fexxv_doppler_shifts_CenX3/Fit_real_line_doppler_shifts.ipynb
+        
+
         
 
     - FUNCTIONS: functions.md
+    
     - CONTRIBUTE: contribute.md
+    
     - REFERENCES: references.md
