+++ paper/paper.md
@@ -49,7 +49,7 @@ The package functions are based on the following key theories:
 
 ### Conservation of Angular Momentum
 
-In systems with non-zero eccentricity, the orbital phase varies non-linearly with observational time due to increased speed at periastron, governed by the conservation of angular momentum. Kepler’s laws further explain that objects sweep out equal areas in equal times (see [@2006imabookC]).
+In systems with non-zero eccentricity, the orbital phase varies non-linearly with observational time due to increased speed at periastron, governed by the conservation of angular momentum. Kepler’s laws further explain that objects sweep out equal areas in equal times (see ---).
 
 $$ r^2 \cdot \omega = h $$
 
@@ -57,19 +57,19 @@ We account for this in all functions and provide tools to convert phase to time 
 
 ### CAK Model
 
-The CAK model [@1975ApJ195157C] describes radiation-driven winds in massive stars, with wind velocity, density, and ionization state varying with distance from the companion star.
+The CAK model --- describes radiation-driven winds in massive stars, with wind velocity, density, and ionization state varying with distance from the companion star.
 
 $$ \rho = \frac{\dot{M}}{4 \pi v R^2} $$
 
-Here, $\rho$  is wind density, $\dot{M}$ is the mass accretion rate, v is orbital speed, and R is the distance to the star. We assume a spherical, un-ionized wind distribution.
+Here, $\rho$  is wind density, $\dot{M}$ is the mass loss rate, v is orbital speed, and R is the distance to the companion. We assume a spherical, un-ionized wind distribution.
 
 ### Accretion Luminosity and Ionization
 
 Accretion powers many close binary systems. The accretion luminosity is given by:
 
-$$ L_{ac} = \frac{GM \dot{M}}{R} $$
+$$ L_{ac} = \frac{G M \dot{M}}{R} $$
 
-where $\L_{ac}$ is the luminosity, $\G$ is the gravitational constant, $\M$ is mass, and $\R$ is the accretion radius. The ionization parameter $\\xi$ is defined as:
+where L$_{ac}$ is the luminosity, G is the gravitational constant, M is the NS mass, and R is the distance to the source. The ionization parameter $\\xi$ is defined as:
 
 $$ \xi = \frac{L_{\rm X}}{n(r_{\rm X}) r_{\rm X}^{2}} $$
 
@@ -85,7 +85,7 @@ $$ \lambda_{D} = \lambda_{\text{rest}}\left(1+\frac{v_{D}}{c}\right) $$
 
 where $\r$ is the orbital radius, $\\omega$ is angular velocity, $\i$ is inclination, and $\\lambda_{\rm D}$ and $\\lambda_{\rm rest}$ are the Doppler-shifted and rest wavelengths, respectively.
 
-Within the fitting functions, we use a particle swarm approach ([@pyswarms], [@10.1162/EVCO_r_00180]) as a classical least squares algorithm does not always converge.
+Within the fitting functions, we use a particle swarm approach (---, ---) as a classical least squares algorithm does not always converge.
 
 ![Some results obtained with the functions contained in this package.](joss.jpg){#sylt width="100%"}
 
@@ -117,7 +117,7 @@ When considering stellar wind (the matter accreted by compact objects), various 
 
 Although these orbital modulations are conceptually simple, they are challenging to analyze. At the same time, they can provide tremendous insights into the orbital mechanics and wind properties of our systems, helping to complete the puzzle when we already have some pieces of information. This is where our tools provide critical support.
 
[…]
