+++ paper/paper.md
@@ -41,7 +41,7 @@ X-ray binaries are tremendously interesting systems. In these extreme environmen
 
 Orbital modulations are observed in nearly all X-ray binary systems. These variations arise from the orbital motions of the system, driven by the relative velocities of the two stars and their changing configurations with respect to each other and the observer.
 
-To aid in the study of these modulations, we introduce `xraybinaryorbit`—a user-friendly Python package designed to simplify the analysis of orbital modulations in X-ray binaries. Whether you are studying the orbital parameters or working to refine your data analysis, this tool is built to help you extract the most from your observations.
+To aid in the study of these modulations, we introduce **xraybinaryorbit** , a user-friendly Python package designed to simplify the analysis of orbital modulations in X-ray binaries. 
 
 # Science Behind
 
@@ -51,7 +51,7 @@ The package functions are based on the following key theories:
 
 In systems with non-zero eccentricity, the orbital phase varies non-linearly with observational time due to increased speed at periastron, governed by the conservation of angular momentum. Kepler’s laws further explain that objects sweep out equal areas in equal times (see ---).
 
-$$ r^2 \cdot \omega = h $$
+$$ r^2 \cdot \omega = Area $$
 
 We account for this in all functions and provide tools to convert phase to time and vice versa.
 
@@ -61,7 +61,7 @@ The CAK model --- describes radiation-driven winds in massive stars, with wind v
 
 $$ \rho = \frac{\dot{M}}{4 \pi v R^2} $$
 
-Here, $\rho$  is wind density, $\dot{M}$ is the mass loss rate, v is orbital speed, and R is the distance to the companion. We assume a spherical, un-ionized wind distribution.
+Here, $\rho$  is wind density, $\dot{M}$ is the mass loss rate, v is the wind velocity and R is the distance to the companion. We assume a spherical, smooth and un-ionized wind distribution.
 
 ### Accretion Luminosity and Ionization
 
@@ -69,7 +69,7 @@ Accretion powers many close binary systems. The accretion luminosity is given by
 
 $$ L_{ac} = \frac{G M \dot{M}}{R} $$
 
-where L$_{ac}$ is the luminosity, G is the gravitational constant, M is the NS mass, and R is the distance to the source. The ionization parameter $\\xi$ is defined as:
+where L$_{ac}$ is the luminosity, G is the gravitational constant, M is the compact object mass, and R is the compact object radius. The ionization parameter $\xi$ is defined as:
 
 $$ \xi = \frac{L_{\rm X}}{n(r_{\rm X}) r_{\rm X}^{2}} $$
 
@@ -79,11 +79,15 @@ We calculate the ionization map within the system plane based on these calculati
 
 The Doppler effect helps determine orbital parameters in X-ray binaries, affecting emission line energies and enabling the analysis of radial velocities.
 
+The Doppler velocity can be calculated as:
+
 $$ v_{D} = (-r\omega \sin\phi \sin i) $$
+
+The doppler length wave affected by Doppler effect be written as:
 
 $$ \lambda_{D} = \lambda_{\text{rest}}\left(1+\frac{v_{D}}{c}\right) $$
 
-where $\r$ is the orbital radius, $\\omega$ is angular velocity, $\i$ is inclination, and $\\lambda_{\rm D}$ and $\\lambda_{\rm rest}$ are the Doppler-shifted and rest wavelengths, respectively.
+where r is the orbital radius, $\omega$ is angular velocity, $\i$ is inclination, and $\lambda_{\rm D}$ and $\\lambda_{\rm rest}$ are the Doppler-shifted and rest wavelengths, respectively.
 
[…]
