+++ paper/paper.bib
@@ -34,7 +34,7 @@ archivePrefix = {arXiv},
 }
 
 
-@INPROCEEDINGS{2016SPIE9905E2FB,
+@INPROCEEDINGS{2016SPIE.9905E..2FB,
   author = {{Barret}, Didier and {Lam Trong}, Thien and {den Herder}, Jan-Willem and {Piro}},
   title = "{The Athena X-ray Integral Field Unit (X-IFU)}",
   keywords = {...},
@@ -72,7 +72,7 @@ archivePrefix = {arXiv},
   year         = {2022},
 }
 
-@ARTICLE{1975ApJ195157C,
+@ARTICLE{1975ApJ...195..157C,
        author = {{Castor}, J.~I. and {Abbott}, D.~C. and {Klein}, R.~I.},
         title = "{Radiation-driven winds in Of stars.}",
       journal = {\apj},
@@ -86,7 +86,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{1982ApJ263835S,
+@ARTICLE{1982ApJ...263..835S,
    author = {{Scargle}, J.~D.},
     title = "{Studies in astronomical time series analysis. II - Statistical aspects of spectral analysis of unevenly spaced data}",
   journal = {\apj},
@@ -122,7 +122,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@BOOK{2006imabookC,
+@BOOK{2006ima..book.....C,
        author = {{Carroll}, Bradley W. and {Ostlie}, Dale A.},
         title = "{An introduction to modern astrophysics and cosmology}",
          year = 2006,
@@ -131,7 +131,7 @@ archivePrefix = {arXiv},
 }
 
 
-@ARTICLE{2022MNRAS512304S,
+@ARTICLE{2022MNRAS.512..304S,
        author = {{Sanjurjo-Ferr{\'\i}n}, G. and {Torrej{\'o}n}, J.~M. and {Rodes-Roca}, J.~J.},
         title = "{The first X-ray spectrum of the high-mass X-ray binary XTE J1855-026 during the compact object eclipse}",
       journal = {\mnras},
