+++ paper/paper.md
@@ -49,7 +49,7 @@ The package functions are based on the following key theories:
 
 ### Conservation of Angular Momentum
 
-In systems with non-zero eccentricity, the orbital phase varies non-linearly with observational time due to increased speed at periastron, governed by the conservation of angular momentum. Kepler’s laws further explain that objects sweep out equal areas in equal times [2006ima..book.....C].
+In systems with non-zero eccentricity, the orbital phase varies non-linearly with observational time due to increased speed at periastron, governed by the conservation of angular momentum. Kepler’s laws further explain that objects sweep out equal areas in equal times [@2006ima..book.....C].
 
 $$ r^2 \cdot \omega = Area $$
 
@@ -57,7 +57,7 @@ We account for this in all functions and provide tools to convert phase to time 
 
 ### CAK Model
 
-The CAK model ([1975ApJ...195..157C]) describes radiation-driven winds in massive stars, with wind velocity, density, and ionization state varying with distance from the companion star.
+The CAK model [@1975ApJ...195..157C] describes radiation-driven winds in massive stars, with wind velocity, density, and ionization state varying with distance from the companion star.
 
 $$ \rho = \frac{\dot{M}}{4 \pi v R^2} $$
 
@@ -65,7 +65,7 @@ Here, $\rho$  is wind density, $\dot{M}$ is the mass loss rate, v is the wind ve
 
 ### Accretion Luminosity and Ionization
 
-Accretion powers many close binary systems @Frank_King_Raine_2002]. The accretion luminosity is given by:
+Accretion powers many close binary systems [@Frank_King_Raine_2002]. The accretion luminosity is given by:
 
 $$ L_{ac} = \frac{G M \dot{M}}{R} $$
 
@@ -101,27 +101,27 @@ The software provides a user-friendly interface for managing the various paramet
 
 ### Fitting Functions
 
-For fitting orbital parameters, the software offers two approaches: least squares (LS) and particle swarm optimization (PSO) [@pyswarms,@10.1162/EVCO_r_00180], denoted by `_ls` and `_ps`, within the functions name respectively. The LS methods is faster but may fail to converge in certain cases, whereas the PS0 is more robust but computationally intensive. Key parameters for PSO include `num_iterations`, `maxiter`, and `swarmsize`. It is recommended to start with smaller values for these parameters (e.g., `num_iterations=3`, `maxiter=100`, and `swarmsize=20`) to evaluate computational demand and adjust accordingly. 
+For fitting orbital parameters, the software offers two approaches: least squares (LS) and particle swarm optimization (PSO) (see [@10.1162/EVCO_r_00180,@pyswarms]), denoted by `_ls` and `_ps`, within the functions name respectively. The LS methods is faster but may fail to converge in certain cases, whereas the PS0 is more robust but computationally intensive. Key parameters for PSO include `num_iterations`, `maxiter`, and `swarmsize`. It is recommended to start with smaller values for these parameters (e.g., `num_iterations=3`, `maxiter=100`, and `swarmsize=20`) to evaluate computational demand and adjust accordingly. 
 
 ### Extended vs. Discrete Fitting
 
-The software provides two fitting methods: `extended` and `discrete`. When fitting data to orbital modulation, a list of `values` corresponding to `time sections` (e.g., phase-resolved spectra or lightcurve segments, wich have a defined time lenght) will be our data to fit. For short-period orbital modulations, which are often sinusoidal, the center of the time section may not properly represent the modulation (that would be a discrete approach, i.e. an discrete time/orbital phase for a discrete value). Instead, it’s necessary to consider the average modulation across the entire `time section` and fit that to the observed `values`.
[…]
