### 2024-10-21 #7220 @xragua
Dear @matteobachetti, I tried to follow your indications both in the documentation pages, the README and in the paper, I hope that now it looks better, let me know what you think, Thanks a lot!
