### 2024-10-21 #7220 @warrickball
@tddesjardins — Have you had a chance to start a review? I noticed you haven't generated a checklist yet.

@matteobachetti — Thanks for your review. You haven't ticked off all the items so I presume you're waiting for further responses from @xragua. Once your satisfied on any given point, please tick off each item, and I suggest also saying you're happy in this comment thread.

@xragua — Note that @matteobachetti has left comments by editing [his review checklist](https://github.com/openjournals/joss-reviews/issues/7220#issuecomment-2348180229). You should address the comments he's raised there.
