### 2024-10-21 #7220 @matteobachetti
@xragua the docs look much better now! Great job!
Small nitpick: there is a typo in the big all-caps title "GUETTING STARTED"
