+++ requirements.txt
@@ -7,3 +7,4 @@ tk
 astropy
 pytest>=6.0
 xraybinaryorbit
+Coverage
