### 2024-10-23 1e5ddda
Update ci-test.yml coverage
