+++ .github/workflows/ci-test.yml
@@ -37,7 +37,7 @@ jobs:
       run: |
         cd test
         coverage run -m pytest --verbose
-        coverage xml -o coverage.xml
+        coverage html
         coverage report
 
     - name: Upload test results
