+++ .github/workflows/ci-test.yml
@@ -52,5 +52,5 @@ jobs:
       if: always()
       uses: actions/upload-artifact@v4  # Updated to v4 for Node.js 20 compatibility
       with:
-        name: coverage-report
+        name: coverage-report-${{ github.run_id }} 
         path: test/coverage.xml  # Upload coverage.xml from 'tests' folder
