+++ paper/paper.bib
@@ -170,3 +170,134 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2022arXiv220205399X},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+
+@ARTICLE{2019ApJ...881...39H,
+       author = {{Huppenkothen}, Daniela and {Bachetti}, Matteo and {Stevens}, Abigail L. and {Migliari}, Simone and {Balm}, Paul and {Hammad}, Omar and {Khan}, Usman Mahmood and {Mishra}, Himanshu and {Rashid}, Haroon and {Sharma}, Swapnil and {Martinez Ribeiro}, Evandro and {Valles Blanco}, Ricardo},
+        title = "{Stingray: A Modern Python Library for Spectral Timing}",
+      journal = {\apj},
+     keywords = {methods: data analysis, methods: statistical, X-rays: binaries, X-rays: general, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - High Energy Astrophysical Phenomena},
+         year = 2019,
+        month = aug,
+       volume = {881},
+       number = {1},
+          eid = {39},
+        pages = {39},
+          doi = {10.3847/1538-4357/ab258d},
+archivePrefix = {arXiv},
+       eprint = {1901.07681},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019ApJ...881...39H},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@software{2018ascl.soft05019B,
+       author = {{Bachetti}, Matteo},
+        title = "{HENDRICS: High ENergy Data Reduction Interface from the Command Shell}",
+ howpublished = {Astrophysics Source Code Library, record ascl:1805.019},
+         year = 2018,
+        month = may,
+          eid = {ascl:1805.019},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2018ascl.soft05019B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+@article{astropy:2013,
+Adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+Adsurl = {http://adsabs.harvard.edu/abs/2013A%26A...558A..33A},
+Archiveprefix = {arXiv},
[…]
