+++ paper/paper.md
@@ -48,7 +48,6 @@ To aid in the study of these modulations, we introduce **xraybinaryorbit** , a u
 The package functions are based on the following key theories:
 
 ### Conservation of Angular Momentum
-
 In systems with non-zero eccentricity, the orbital phase varies non-linearly with observational time due to increased speed at periastron, governed by the conservation of angular momentum. Kepler’s laws further explain that objects sweep out equal areas in equal times [@2006ima..book.....C].
 
 $$ r^2 \cdot \omega = Area $$
@@ -56,15 +55,13 @@ $$ r^2 \cdot \omega = Area $$
 We account for this in all functions and provide tools to convert phase to time and vice versa.
 
 ### CAK Model
-
-The CAK model [@1975ApJ...195..157C] describes radiation-driven winds in massive stars, with wind velocity, density, and ionization state varying with distance from the companion star.
+The CAK model [@1975ApJ...195..157C] describes radiation-driven winds in massive stars, with wind velocity, density, and ionization state varying with distance to the companion star.
 
 $$ \rho = \frac{\dot{M}}{4 \pi v R^2} $$
 
 Here, $\rho$  is wind density, $\dot{M}$ is the mass loss rate, v is the wind velocity and R is the distance to the companion. We assume a spherical, smooth and un-ionized wind distribution.
 
 ### Accretion Luminosity and Ionization
-
 Accretion powers many close binary systems [@Frank_King_Raine_2002]. The accretion luminosity is given by:
 
 $$ L_{ac} = \frac{G M \dot{M}}{R} $$
@@ -73,17 +70,14 @@ where L$_{ac}$ is the luminosity, G is the gravitational constant, M is the comp
 
 $$ \xi = \frac{L_{\rm X}}{n(r_{\rm X}) r_{\rm X}^{2}} $$
 
-We calculate the ionization map within the system plane based on these calculations.
-
 ### Doppler Effect
-
-The Doppler effect helps determine orbital parameters in X-ray binaries, affecting emission line energies and enabling the analysis of radial velocities.
+The Doppler effect refers to the change in frequency or wavelength of a wave (such as sound or light) in relation to an observer moving relative to the source. 
 
 The Doppler velocity provoked by orbital motion can be calculated as:
 
 $$ v_{D} = (-r\omega \sin\phi \sin i) $$
 
-The doppler length wave affected by Doppler effect be written as:
+The observed Doppler length wave can be written as:
 
 $$ \lambda_{D} = \lambda_{\text{rest}}\left(1+\frac{v_{D}}{c}\right) $$
 
@@ -94,8 +88,7 @@ where r is the orbital radius, $\omega$ is angular velocity, $\i$ is inclination
 
 ## Similar software
 
-Other software packages that can be applied to X-ray binary system analysis include Stingray [2019ApJ...881...39H] and HENDRICS [2018ascl.soft05019B], which focus on timing analysis of astronomical data. Jaxspec specializes in spectral analysis using Bayesian inference techniques. On the other hand, Astropy [astropy:2022] is a comprehensive package for general astronomy computations and data handling. In contrast, xraybinaryorbit is highly specialized, providing detailed insights into orbital modulations within X-ray binary systems. Each of these packages offers distinct tools, complementing one another based on the focus—whether it's timing, spectral, or orbital analysis.
-
[…]
