+++ paper/paper.bib
@@ -65,7 +65,6 @@ archivePrefix = {arXiv},
     doi = {10.1038/s41592-019-0686-2},
 }
 
-
 @misc{pyswarms,
   title        = {{PySwarms}: A Python-Based Swarm Optimization Library},
   howpublished = {\url{https://pythonhosted.org/pyswarm/}},
