+++ paper/paper.bib
@@ -71,7 +71,7 @@ archivePrefix = {arXiv},
   year         = {2022},
 }
 
-@ARTICLE{1975ApJ...195..157C,
+@ARTICLE{1975ApJ195157C,
        author = {{Castor}, J.~I. and {Abbott}, D.~C. and {Klein}, R.~I.},
         title = "{Radiation-driven winds in Of stars.}",
       journal = {\apj},
@@ -121,7 +121,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@BOOK{2006ima..book.....C,
+@BOOK{2006imabookC,
        author = {{Carroll}, Bradley W. and {Ostlie}, Dale A.},
         title = "{An introduction to modern astrophysics and cosmology}",
          year = 2006,
@@ -171,7 +171,7 @@ archivePrefix = {arXiv},
 }
 
 
-@ARTICLE{2019ApJ...881...39H,
+@ARTICLE{2019ApJ88139H,
        author = {{Huppenkothen}, Daniela and {Bachetti}, Matteo and {Stevens}, Abigail L. and {Migliari}, Simone and {Balm}, Paul and {Hammad}, Omar and {Khan}, Usman Mahmood and {Mishra}, Himanshu and {Rashid}, Haroon and {Sharma}, Swapnil and {Martinez Ribeiro}, Evandro and {Valles Blanco}, Ricardo},
         title = "{Stingray: A Modern Python Library for Spectral Timing}",
       journal = {\apj},
@@ -300,7 +300,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@article{refId0,
+@article{newcenx3,
 	author = {{Sanjurjo-Ferrín, G.} and {Torrejón, J. M.} and {Oskinova, L.} and {Postnov, K.} and {Rodes-Roca, J. J.} and {Schulz, N.} and {Nowak, M.}},
 	title = {Revealing the state transition of Cen X-3 at high spectral resolution with Chandra
 },
