### 2024-10-23 #7220 @xragua
@matteobachetti, @tddesjardins and @warrickball, I have made the following updates to the project:
*Modified the paper, pages, and README files.
*Added coverage calculation in the CI test.
*Included a few lines in the community guidelines.
*I look forward to your feedback. Thank you!
