### 2024-10-24 #7220 @matteobachetti
@xragua: thanks for all the work you are doing to address my comments.
Besides all the nitpicks I'm doing through Issues, I do have one general suggestion that I make in this context though: all the testing/coverage apparatus I've been recommending is incredibly useful, but it's just esthetics (cute green ticks in a checklist) if you commit to main without checking them. Developers of robust open source projects work in branches/forks, test locally, and when code is ready, they do a pull request, during which tests are automatically executed by Github Actions or similar systems on multiple python versions and architectures, coverage is updated, and so on. It seems like a needless burden, but it's very effective in finding code breaks before the code goes into production.
