### 2024-10-24 a04a381
Update ci-test.yml
