+++ .github/workflows/ci-test.yml
@@ -28,8 +28,25 @@ jobs:
       with:
         python-version: ${{ matrix.python-version }}
 
-    - name: Install dependencies
+     - name: Install dependencies
       run: |
         pip install --upgrade pip
         pip install --upgrade xraybinaryorbit
+        pip install pytest
+        pip install pytest-cov
 
+    - name: Run tests with coverage
+      run: |
+        cd test  # Navigate to the test directory
+        pytest --cov=./ --cov-report=xml  # Generate the coverage report
+
+    - name: Upload coverage to Codecov
+      uses: codecov/codecov-action@v4
+      with:
+        file: ./coverage.xml  
+        env_vars: OS,PYTHON
+        fail_ci_if_error: true
+        flags: unittests
+        name: codecov-umbrella
+        token: ${{ secrets.CODECOV_TOKEN }}  
+        verbose: true
