+++ .github/workflows/ci-test.yml
@@ -28,7 +28,7 @@ jobs:
       with:
         python-version: ${{ matrix.python-version }}
 
-     - name: Install dependencies
+    - name: Install dependencies
       run: |
         pip install --upgrade pip
         pip install --upgrade xraybinaryorbit
@@ -43,10 +43,11 @@ jobs:
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v4
       with:
-        file: ./coverage.xml  
+        file: ./coverage.xml  # Ensure this matches the actual coverage report file
         env_vars: OS,PYTHON
         fail_ci_if_error: true
         flags: unittests
         name: codecov-umbrella
-        token: ${{ secrets.CODECOV_TOKEN }}  
+        token: ${{ secrets.CODECOV_TOKEN }}  # Ensure you have this secret set in your repository
         verbose: true
+
