+++ .github/workflows/ci-test.yml
@@ -43,11 +43,11 @@ jobs:
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v4
       with:
-        file: ./coverage.xml  # Ensure this matches the actual coverage report file
+        file: ./coverage.xml  
         env_vars: OS,PYTHON
         fail_ci_if_error: true
         flags: unittests
         name: codecov-umbrella
-        token: ${{ secrets.CODECOV_TOKEN }}  # Ensure you have this secret set in your repository
+        token: ${{ secrets.CODECOV_TOKEN }}  
         verbose: true
 
