### 2024-10-24 dfdf86c
Delete my-project/docs directory
