+++ my-project/docs/notebooks/example_doppler_theoretical/spiral_in_orbit.txt
@@ -1,2 +1,0 @@
-iphase,semimajor,orbitalperiod,eccentricity,periapsis,inclination,iphase_spiral,semimajor_spiral,b,omega,inclination_spiral,Rstar,Mstar1,Mstar2,wind_vel,feature
-0.0,1.0,2.0,0.2,170.0,60.0,0.0,1.0,0.001,0.0001,60.0,12.0,20.0,2.0,0.0,6.7
