### 2024-10-24 e4f8b81
Delete my-project/mkdocs.yml
