+++ my-project/mkdocs.yml
@@ -1,0 +1,44 @@
+site_name: xraybinaryorbit
+
+theme:
+    palette:
+        scheme: slate
+        primary: pink
+        
+    name: readthedocs
+    logo: images/logo.png  # Add this line to specify the logo path
+
+extra_css:
+    - css/custom.css  # Include your custom CSS file here
+
+plugins:
+    - mkdocs-jupyter
+
+nav:
+    -  GETTING STARTED: index.md
+    
+    - EXAMPLES:
+    
+
+        - Time Phase Relation Theoretical: notebooks/example_time_phase_relation_theoretical/time_phase_relation.ipynb
+    
+        - Doppler Theoretical in four scenarios: notebooks/example_doppler_theoretical/Doppler_theoretical_functions.ipynb
+        
+        - Density throuth the orbit line and NH Theoretical: notebooks/example_density_NH_theoretical/Wind_density_and_NH.ipynb
+        
+        - Density and ionization through the path of the emitted radiation depending on the orbital Phase: notebooks/example_density_ionization_orbital_phase/density_and_chi_orbital_phase.ipynb
+        
+        - Ionization Map of 4U0114+65: notebooks/example_ionization_map_4U0114+65/Ionization_map.ipynb
+        
+        - NS Spin Shape Vela X-1: notebooks/example_NS_spin_shape_VelaX1/Pulse_shape_Velax1.ipynb
+
+        - Fe XXV Doppler Shifts Cen X-3: notebooks/example_fexxv_doppler_shifts_CenX3/Fit_real_line_doppler_shifts.ipynb
+        
+
+        
+
+    - FUNCTIONS: functions.md
+    
+    - CONTRIBUTE: contribute.md
+    
+    - REFERENCES: references.md
