### 2024-10-24 bb77158
Delete my-project/site directory
