+++ my-project/site/notebooks/example_time_phase_relation_theoretical/phase_time.txt
@@ -1,2 +1,0 @@
-iphase,semimajor,orbitalperiod,eccentricity,periapsis,Rstar,Mstar1,Mstar2
-0.0,1.5,2.0,0.2,200.0,12.0,22.0,1.4
