+++ my-project/docs/contribute.md
@@ -1,0 +1,18 @@
+
+# Contributing to XRAYBINARYORBIT
+
+Thank you for taking the time to contribute! We appreciate your effort in helping to improve the **xraybinaryorbit** package.
+
+## How to Contribute
+
+### 1. Reporting Bugs
+
+If you find a bug, please open an issue [here](https://github.com/xragua/XRAYBINARYORBIT/issues)
+
+### 2. Suggesting Enhancements
+
+Have a feature request or improvement idea? Please open a feature request [here](https://github.com/xragua/XRAYBINARYORBIT/issues) 
+
+### 3. Contribute to the package
+We welcome contributions of all kinds! Whether it's fixing a simple typo in our documentation or developing a significant new feature, every contribution is valuable and appreciated. 
+We welcome contributors who will abide by the [Python Software Foundation Code of Conduct.](https://policies.python.org/python.org/code-of-conduct/)
