+++ my-project/docs/references.md
@@ -1,0 +1,49 @@
+## REFERENCES
+
+1. Bonyadi, M. R., & Michalewicz, Z. (2017). Particle Swarm Optimization for Single Objective Continuous Space Problems: A Review. *Evolutionary Computation, 25*(1), 1-54. [https://doi.org/10.1162/EVCO_r_00180](https://doi.org/10.1162/EVCO_r_00180)
+
+
+
+2. Sanjurjo-Ferrín, G., Torrejón, J. M., Postnov, K., Oskinova, L., Rodes-Roca, J. J., & Bernabeu, G. (2021). X-ray variability of the HMXB Cen X-3: evidence for inhomogeneous accretion flows. *Monthly Notices of the Royal Astronomical Society, 501*(4), 5892-5909. [https://doi.org/10.1093/mnras/staa3953](https://doi.org/10.1093/mnras/staa3953)
+
+
+
+3. Barret, D., Lam Trong, T., den Herder, J.-W., & Piro. (2016). The Athena X-ray Integral Field Unit (X-IFU). In J.-W. A. den Herder, T. Takahashi, & M. Bautz (Eds.), *Space Telescopes and Instrumentation 2016: Ultraviolet to Gamma Ray* (Vol. 9905, p. 99052F). Society of Photo-Optical Instrumentation Engineers (SPIE). [https://doi.org/10.1117/12.2232432](https://doi.org/10.1117/12.2232432)
+
+
+
+4. Virtanen, P., Gommers, R., Oliphant, T. E., Haberland, M., Reddy, T., Cournapeau, D., Burovski, E., Peterson, P., Weckesser, W., Bright, J., van der Walt, S. J., Brett, M., Wilson, J., Jarrod Millman, K., Mayorov, N., Nelson, A. R. J., Jones, E., Kern, R., Larson, E., Carey, C. J., Polat, İ., Feng, Y., Moore, E. W., VanderPlas, J., Laxalde, D., Perktold, J., Cimrman, R., Henriksen, I., Quintero, E. A., Harris, C. R., Archibald, A. M., Ribeiro, A. H., Pedregosa, F., van Mulbregt, P., & Contributors, S. (2020). SciPy 1.0: Fundamental Algorithms for Scientific Computing in Python. *Nature Methods, 17*, 261-272. [https://doi.org/10.1038/s41592-019-0686-2](https://doi.org/10.1038/s41592-019-0686-2)
+
+
+
+5. PySwarms: A Python-Based Swarm Optimization Library. (2022). Retrieved from [https://pythonhosted.org/pyswarm/](https://pythonhosted.org/pyswarm/)
+
+
+
+6. Castor, J. I., Abbott, D. C., & Klein, R. I. (1975). Radiation-driven winds in Of stars. *The Astrophysical Journal, 195*, 157-174. [https://doi.org/10.1086/153315](https://doi.org/10.1086/153315)
+
+
+
+7. Scargle, J. D. (1982). Studies in astronomical time series analysis. II - Statistical aspects of spectral analysis of unevenly spaced data. *The Astrophysical Journal, 263*, 835-853. [https://doi.org/10.1086/160554](https://doi.org/10.1086/160554)
+
+
+
+8. Frank, J., King, A., & Raine, D. (2002). *Accretion Power in Astrophysics* (3rd ed.). Cambridge University Press.
+
+
+
[…]
