+++ my-project/docs/notebooks/example_fexxv_doppler_shifts_CenX3/bounds_disc.txt
@@ -1,0 +1,3 @@
+iphase,semimajor,orbitalperiod,eccentricity,periapsis,inclination,Rstar,Mstar1,Mstar2,iphase2,semimajor2,orbitalperiod2,eccentricity2,periapsis2,inclination2,Mass3,feature,wind_vel
+0.18,1.5,2.087,0.0,0.0,67.0,11.8,18.7,1.33,0.0,0.5,0.14,0.0,0.0,67.0,1e-60,6.63,0.0
+0.25,1.7,2.09,1e-05,1e-70,76.7,12.2,22.77,1.45,0.99,1.0,0.16,0.46,360.0,76.7,1e-22,6.7,1e-70
