+++ my-project/docs/notebooks/example_doppler_theoretical/orbit.txt
@@ -1,0 +1,2 @@
+iphase,semimajor,orbitalperiod,eccentricity,periapsis,inclination,Rstar,Mstar1,Mstar2,wind_vel,feature
+0.0,1.5,2.0,0.2,200.0,90.0,12.0,22.0,1.4,600.0,6.7
