### 2024-11-11 cdde486
Updated JOSS paper draft with improved text and corrections
