+++ paper/paper.bib
@@ -34,7 +34,7 @@ archivePrefix = {arXiv},
 }
 
 
-@INPROCEEDINGS{2016SPIE.9905E..2FB,
+@INPROCEEDINGS{2016SPIE.9905E.2FB,
   author = {{Barret}, Didier and {Lam Trong}, Thien and {den Herder}, Jan-Willem and {Piro}},
   title = "{The Athena X-ray Integral Field Unit (X-IFU)}",
   keywords = {...},
@@ -130,7 +130,7 @@ archivePrefix = {arXiv},
 }
 
 
-@ARTICLE{2022MNRAS.512..304S,
+@ARTICLE{2022MNRAS.512.304S,
        author = {{Sanjurjo-Ferr{\'\i}n}, G. and {Torrej{\'o}n}, J.~M. and {Rodes-Roca}, J.~J.},
         title = "{The first X-ray spectrum of the high-mass X-ray binary XTE J1855-026 during the compact object eclipse}",
       journal = {\mnras},
