+++ paper/paper.md
@@ -1,4 +1,4 @@
----
+a---
 title: 'Orbital Dynamics in X-RAY stellar binary systems'
 tags:
   - Python
@@ -37,18 +37,20 @@ bibliography: paper.bib
 
 X-ray astronomy is a young discipline, spanning no more than a few decades. The Earth's atmosphere is opaque to this type of radiation, so observations in this part of the spectrum were delayed until the space era began, when rocket launchers carrying X-ray telescopes revealed the universe from a brand-new point of view.
 
-X-ray binaries are tremendously interesting systems. In these extreme environments, a compact object—either a neutron star or black hole—draws in matter from a companion star, producing intense X-ray emissions. These systems offer a unique window into extreme physics, from the effects of strong gravity and relativistic jets to the presence of intense magnetic fields.
+X-ray binaries are tremendously interesting systems. In these extreme environments, a compact object (either a neutron star, a black hole or a white dwarf) draws in matter from a companion star, producing X-ray radiation. These systems offer a unique window into extreme physics, from the effects of strong gravity and relativistic jets to the presence of intense magnetic fields.
 
-Orbital modulations are observed in nearly all X-ray binary systems. These variations arise from the orbital motions of the system, driven by the relative velocities of the two stars and their changing configurations with respect to each other and the observer.
+Orbital modulations are observed in nearly all X-ray binary systems. These variations arise from the orbital motions within the system, driven by the relative velocities of the two stars and/or their changing configurations with respect to each other and the observer.
 
-To aid in the study of these modulations, we introduce **xraybinaryorbit**, a user-friendly Python package designed to simplify the analysis of orbital modulations in X-ray binaries.
-There are three groups of funtions within this package, theoretical, fitting and timing. The software provides a user-friendly interface for managing the various parameters that influence orbital modulations. Upon first use, the user inputs parameters through a form, which are then saved in a file within the running directory, this file is automatically loaded in future runs, eliminating the need to re-enter all parameters. The software offers two fitting approaches: least squares (LS) and particle swarm optimization (PSO) [@10.1162/EVCO_r_00180,@pyswarms]. The LS methods is faster but may fail to converge in certain cases, whereas the PS0 is more robust but computationally intensive.  Some results obtained with this packages are shown in Figure 1. 
+To aid in the study of these modulations, we introduce xraybinaryorbit, a user-friendly Python package designed to simplify the analysis of orbital modulations in X-ray binaries. The package is organized into three groups of functions: theoretical, fitting, and timing. It features an intuitive interface for managing the various parameters that influence orbital modulations. Upon first use, the user inputs parameters via a form; these parameters are then saved to a file within the working directory. This file is automatically loaded in future sessions, eliminating the need to re-enter parameters each time.
 
-There are many software that can be applied to X-ray binary system analysis, including **Stingray** [@2019ApJ88139H] which focus on timing analysis of astronomical data, **Jaxspec** which specializes in spectral fitting using Bayesian inference. **Lightkurve**,  facilitates the analysis of time-series data from space missions like Kepler and TESS. 
-A well know comprehensive package for general astronomy computations and data handling is **Astropy** [@astropy:2022]. 
-In contrast **xraybinaryorbit** is highly specialized, providing detailed insights into orbital modulations within X-ray binary systems. 
[…]
