### 2024-11-11 3a2907b
Updated JOSS paper draft with improved text and corrections
