+++ paper/paper.md
@@ -27,7 +27,7 @@ affiliations:
  - name: Instituto Universitario de Física Aplicada a las Ciencias y las Tecnologías, Universidad de Alicante, 03690 Alicante, Spain
    index: 1
 
-date: 21 March 2024
+date: 11 November 2024
 bibliography: paper.bib
 
 ---
