### 2024-11-11 378a6e4
Updated JOSS paper draft with improved text and corrections
