+++ paper/paper.md
@@ -1,4 +1,4 @@
-a---
+---
 title: 'Orbital Dynamics in X-RAY stellar binary systems'
 tags:
   - Python
