### 2024-11-11 1269c3e
Updated JOSS paper draft with improved text and corrections
