### 2024-11-12 #7220 @matteobachetti
@warrickball @xragua I just realized I had not followed up here. The paper has improved a lot, and I have no major comments left, just one of personal taste: for what I could interpret of the statement of need, I think it should just contain the paragraph about the other software packages, while what is currently in the statement of need would fit perfectly in the general description. But it's a matter of personal taste. Thanks @xragua for addressing my comments.
