+++ paper/paper.md
@@ -39,13 +39,21 @@ X-ray astronomy is a young discipline, spanning no more than a few decades. The 
 
 X-ray binaries are tremendously interesting systems. In these extreme environments, a compact object (either a neutron star, a black hole or a white dwarf) draws in matter from a companion star, producing X-ray radiation. These systems offer a unique window into extreme physics, from the effects of strong gravity and relativistic jets to the presence of intense magnetic fields.
 
-Orbital modulations are observed in nearly all X-ray binary systems. These variations arise from the orbital motions within the system, driven by the relative velocities of the two stars and/or their changing configurations with respect to each other and the observer.
+Orbital modulations are observed in nearly all X-ray binary systems. These variations arise from the orbital motions within the system, driven by the relative velocities of the two stars and/or their changing configurations with respect to each other and the observer. As an example, we might detect slight oscillating shifts in the center of emission lines and/or small variations in a neutron star's spin period. This phenomenon is likely caused by the Doppler effect. Our code helps translate this data into the orbital parameters responsible for such Doppler shifts.
+
+Although the process seems simple, it becomes complex due to factors like inclination, eccentricity, periapsis, and the masses of the stars involved. If the eccentricity is greater than zero, the velocity around the orbit is not constant, making the analysis even more intricate.
+
+When considering stellar wind (the matter accreted by compact objects), various combinations of these factors result in orbital modulations. Eccentricity causes changes in wind density throughout the orbit, affecting the amount of accreted matter. Both eccentricity and inclination influence the absorption column faced by the emitted radiation and the ionization of the stellar wind, which also varies with orbital phase.
+
+Although these orbital modulations are conceptually simple, they are challenging to analyze. At the same time, they can provide tremendous insights into the orbital mechanics and wind properties of our systems, may help to complete the puzzle when we already have some pieces of information or to give us a list of possible scenarios compatible with our data. This is where our tools provide valuable support.
+
+Historically, the primary limitation in this type of analysis has been the lack of resolution for detailed phase-resolved observations. However, upcoming high-resolution missions like XRISM [@2022arXiv220205399X] and New Athena [@2016SPIE.9905E.2FB] promise to significantly improve the quality of these analyses. In addition to better resolution, advances in computational power have been crucial. Many of these tools have already been successfully applied to studies using XMM-Newton and Chandra data, enabling analyses that were previously impossible. [2022MNRAS.512.304S;@2021MNRAS.501.5892S;@newcenx3]
 
 To aid in the study of these modulations, we introduce xraybinaryorbit, a user-friendly Python package designed to simplify the analysis of orbital modulations in X-ray binaries. The package is organized into three groups of functions: theoretical, fitting, and timing. It features an intuitive interface for managing the various parameters that influence orbital modulations. Upon first use, the user inputs parameters via a form; these parameters are then saved to a file within the working directory. This file is automatically loaded in future sessions, eliminating the need to re-enter parameters each time.
 
 The software offers two fitting methods: least squares (LS) and particle swarm optimization (PSO) [@10.1162/EVCO_r_00180, @pyswarms]. The LS method is faster but may fail to converge in some cases, whereas PSO is more robust but computationally intensive. Some results obtained with this package are shown in Figure 1.
 
[…]
