+++ paper/paper.md
@@ -39,17 +39,17 @@ X-ray astronomy is a young discipline, spanning no more than a few decades. The 
 
 X-ray binaries are tremendously interesting systems. In these extreme environments, a compact object (either a neutron star, a black hole or a white dwarf) draws in matter from a companion star, producing X-ray radiation. These systems offer a unique window into extreme physics, from the effects of strong gravity and relativistic jets to the presence of intense magnetic fields.
 
-Orbital modulations are observed in nearly all X-ray binary systems. These variations arise from the orbital motions within the system, driven by the relative velocities of the two stars and/or their changing configurations with respect to each other and the observer. As an example, we might detect slight oscillating shifts in the center of emission lines and/or small variations in a neutron star's spin period. This phenomenon is likely caused by the Doppler effect. Our code helps translate this data into the orbital parameters responsible for such Doppler shifts.
+Orbital modulations are observed in nearly all X-ray binary systems. These variations arise from the orbital motions within the system, driven by the relative velocities of the two stars and/or their changing configurations with respect to each other and the observer. 
 
-Although the process seems simple, it becomes complex due to factors like inclination, eccentricity, periapsis, and the masses of the stars involved. If the eccentricity is greater than zero, the velocity around the orbit is not constant, making the analysis even more intricate.
+For instance, we might detect slight oscillating shifts in the center of emission lines and/or small variations in a neutron star's spin period. This phenomenon is likely caused by the Doppler effect.Our code translates this data into orbital parameters that explain these Doppler shifts. Although the process seems simple, it becomes complex due to factors like inclination, eccentricity, periapsis, and the masses of the stars involved. If the eccentricity is greater than zero, the velocity around the orbit is not constant, making the analysis even more intricate.
 
 When considering stellar wind (the matter accreted by compact objects), various combinations of these factors result in orbital modulations. Eccentricity causes changes in wind density throughout the orbit, affecting the amount of accreted matter. Both eccentricity and inclination influence the absorption column faced by the emitted radiation and the ionization of the stellar wind, which also varies with orbital phase.
 
 Although these orbital modulations are conceptually simple, they are challenging to analyze. At the same time, they can provide tremendous insights into the orbital mechanics and wind properties of our systems, may help to complete the puzzle when we already have some pieces of information or to give us a list of possible scenarios compatible with our data. This is where our tools provide valuable support.
 
-Historically, the primary limitation in this type of analysis has been the lack of resolution for detailed phase-resolved observations. However, upcoming high-resolution missions like XRISM [@2022arXiv220205399X] and New Athena [@2016SPIE.9905E.2FB] promise to significantly improve the quality of these analyses. In addition to better resolution, advances in computational power have been crucial. Many of these tools have already been successfully applied to studies using XMM-Newton and Chandra data, enabling analyses that were previously impossible. [2022MNRAS.512.304S;@2021MNRAS.501.5892S;@newcenx3]
+Historically, the primary limitation in this type of analysis has been the lack of resolution for detailed phase-resolved observations. However, upcoming high-resolution missions like XRISM [@2022arXiv220205399X] and New Athena [@2016SPIE.9905E.2FB] promise to significantly improve the quality of these analyses. In addition to better resolution, advances in computational power have been crucial. Many of these tools have already been successfully applied to studies using XMM-Newton and Chandra data, enabling analyses that were previously impossible. [@2022MNRAS.512.304S;@2021MNRAS.501.5892S;@newcenx3]
 
[…]
