+++ paper/paper.bib
@@ -301,7 +301,7 @@ archivePrefix = {arXiv},
 }
 
 @article{newcenx3,
-	author = {{Sanjurjo-Ferrín, G.} and {Torrejón, J. M.} and {Oskinova, L.} and {Postnov, K.} and {Rodes-Roca, J. J.} and {Schulz, N.} and {Nowak, M.}},
+	author = {{Sanjurjo-Ferrín, G.} , {Torrejón, J. M.} ,{Oskinova, L.} , {Postnov, K.} ,{Rodes-Roca, J. J.} , {Schulz, N.} , {Nowak, M.}},
 	title = {Revealing the state transition of Cen X-3 at high spectral resolution with Chandra
 },
 	DOI= "10.1051/0004-6361/202449724",
