+++ paper/paper.bib
@@ -16,7 +16,7 @@
 
 
 @ARTICLE{2021MNRAS.501.5892S,
-       author = {{Sanjurjo-Ferr{\'\i}n}, G. and {Torrej{\'o}n}, J.~M. and {Postnov}, K. and {Oskinova}, L. and {Rodes-Roca}, J.~J. and {Bernabeu}, G.},
+       author = {{Sanjurjo-Ferr{\'\i}n}, G. and {Torrej{\'o}n}, J.~M. and {Postnov}, K. and {Oskinova}, L., {Rodes-Roca}, J.~J. and {Bernabeu}, G.},
         title = "{X-ray variability of the HMXB Cen X-3: evidence for inhomogeneous accretion flows}",
       journal = {\mnras},
      keywords = {accretion, accretion discs, eclipses, pulsars: individual: Cen X-3, X-rays: binaries, Astrophysics - High Energy Astrophysical Phenomena},
