### 2024-11-12 #7220 @xragua
@matteobachetti, I changed  a little bit the structure of the paper as suggested, I think that now looks better and makes more sense. Thank you very much!
