### 2024-11-13 #7220 @warrickball
@matteobachetti — Thanks! If that concludes your review, would you mind ticking off the remaining items in your [review checklist](https://github.com/openjournals/joss-reviews/issues/7220#issuecomment-2348180229), that'll help make it clear that you've looked at all the items. Thanks!
