### 2024-11-13 #7220 @warrickball
Hi @renecotyfanboy, and thanks again for agreeing to review. This is the review thread for the paper. All of our correspondence will now happen here.

Please read the "Reviewer instructions & questions" in the [first comment](https://github.com/openjournals/joss-reviews/issues/7220#issue-2523224200) above, and generate your checklist by commenting `@editorialbot generate my checklist` on this issue. As you go over the submission, please check off any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.

The JOSS review is different from most other journals. We aim to work with the authors to help them meet our criteria instead of merely passing judgement on the submission. We also encourage reviewers to submit issues and pull requests on the software repository. When doing so, please mention `openjournals/joss-reviews#7220` so that the issue/PR is linked to this thread. Please also feel free to comment and ask questions on this thread.

This paper has been delayed by a reviewer having to withdraw so it'd be great if you could start as soon as you can. JOSS reviews are iterative—rather than monolithic—and the authors can start responding to any comments you raise while you continue to review other parts of the submission.

Don't hesitate to ask any questions you might have about the process.
