### 2024-11-18 xragua/xraybinaryorbit#14 @renecotyfanboy
[MTN] Duplicate License files
Hi @GracielaSanjurjoFerrin
I am slowly starting the review of this package, here is my first comment!
The LICENSE file seems to be duplicated (LICENSE & LICENSE.txt)
