### 2024-11-18 xragua/xraybinaryorbit#15 @renecotyfanboy
[DOC] Provide default values for the examples
It would be great to provide the values you used for the plots in this example
