### 2024-11-18 xragua/xraybinaryorbit#15 @renecotyfanboy
I noticed that the first example was running for nothing because I didn't have interactive stuff on, so this applies also for all the examples I guess
