### 2024-11-18 #7220 @renecotyfanboy
I am not sure where to comment this, but my review was pretty fast since a lot of work has already been put into this repo. I opened two very minor issues (xragua/xraybinaryorbit#14, xragua/xraybinaryorbit#15) and will wait for the second one to run the examples on my computer.

Also, I see you reference `jaxspec` in your paper, we now have [a proper reference](https://ui.adsabs.harvard.edu/abs/2024A%26A...690A.317D/abstract) 😉

Now waiting for @xragua

**Some very broad/general extra suggestions that are not blocking**

- You should split the repository into various submodules, having everything in a big file is a nightmare for people that might be interested in contributing to the source code
- I find the interactivity a bit weird. This makes your soft "hard-to-script" and to integrate in other softwares
- This may be cool to expose the fit likelihood for external user, to plug other optimizers/metaheuristics/samplers to solve the fitting problem
- Maybe consider using `JAX` 😄 I'll try not to be too much of a proselytiser there, but JAX can both speed up your already `numpy` friendly code and you could gain so much from the autograd feature.

These are all suggestions, do whatever you want with this, there's no obligation whatsoever.
