### 2024-11-19 #7220 @xragua
@renecotyfanboy thank you so much for agreeing to review this project! :)

I’m currently working on the corrections, though progress has been a bit slower than I would like, as I’m attending a workshop this week. Nevertheless, I will try to address at least all the issues within the next few days.

Regarding the examples: the input parameters are stored in a .txt file in the same folder as the Jupyter notebooks, which are read directly when the function is set to load_directly=True. If load_directly=False, the parameters are uploaded through the form. If you download this .txt file to the same directory as the notebooks and  set load_directly=True everything should work well. We chose this approach because the number of required parameters is quite large, and a form-like input makes it more intuitive and easier to experiment and play with different values. However, I understand that this may not always be the most convenient approach. I'm currently adding extra functionality so the parameters can also be passed directly as a list and displayed for easy verification.

Thank you very much for all the advices! This is quite new to me and I really appreciate all the guidance. I will write again when I get to finish with the new functionallity.
