+++ xraybinaryorbit/xraybinaryorbit.py
@@ -513,16 +513,11 @@ def _load_values_to_interface(param_list, fixed_values, name):
 
     root.mainloop()
     
-
-def _manage_parameters(param_list, name, load_directly=False):
+def _manage_parameters(param_list, name, load_directly=False, parameter_list=None):
     """
     Manage parameter values by loading from a file, displaying a user input form for modification,
-    and saving the updated values back to the file.
-
-    This function first attempts to load previously saved parameter values from a file named `name.txt`.
-    If the file is not found, it initializes the values to `NaN`. After loading the values,
-    it displays a graphical user interface (GUI) using the `_load_values_to_interface` function
-    to allow the user to update the parameter values. Finally, it saves the updated values back to the same file.
+    and saving the updated values back to the file. This version also allows loading parameters
+    from a given array.
 
     Parameters
     ----------
@@ -532,28 +527,68 @@ def _manage_parameters(param_list, name, load_directly=False):
     name : str
         The base name for the text file where the parameters will be saved and loaded from.
         The file is expected to be named `{name}.txt`.
+    load_directly : bool, optional
+        If True, loads parameters directly from the file without displaying the GUI.
+    parameter_list : list of float, optional
+        If provided, this array of values will be used as the parameter values instead of loading from a file,
+        and no form will be displayed.
 
     Returns
     -------
     fixed_values_list : list of float
-        A list of the final parameter values, either loaded from the file or updated through user input.
+        A list of the final parameter values, either loaded from the file, provided as an array,
+        or updated through user input.
     """
 
     global fixed_values_list
 
-    # Try to load previous fixed values if available
+    # Initialize the flag to ensure it is always defined
+    should_print_after_modification = False
+
+    # If parameter_list is provided, load values directly and skip the form
+    if parameter_list is not None:
+        fixed_values_list = parameter_list
+        print("Loaded parameters from the provided array:")
+        for param, value in zip(param_list, fixed_values_list):
+            print(f"{param}: {value}")
+
+        # Save the provided parameters directly to the file
+        with open(f"{name}.txt", "w") as file:
+            file.write(",".join(map(str, param_list)) + "\n")  # Write parameter names
+            file.write(",".join(map(str, fixed_values_list)) + "\n")  # Write fixed values
+
+        # Skip displaying the form
+        return fixed_values_list
+
+    # If parameter_list is not provided, proceed with loading from the file or showing the form
     file_exists = True
     try:
         with open(f"{name}.txt", "r") as file:
             lines = file.readlines()
             fixed_values_list = [float(val) for val in lines[1].strip().split(",")]
+            if load_directly:
+                print("Loaded parameters from file:")
+                for param, value in zip(param_list, fixed_values_list):
+                    print(f"{param}: {value}")
+            else:
+                should_print_after_modification = True
     except FileNotFoundError:
         file_exists = False
         fixed_values_list = [np.nan] * len(param_list)
+        print("File not found. Initializing parameters to NaN:")
+        for param in param_list:
+            print(f"{param}: NaN")
 
-    # If load_directly is True but file does not exist, still display the GUI
+    # If load_directly is False or the file does not exist, display the GUI
     if not load_directly or not file_exists:
[…]
