+++ xraybinaryorbit/helpers/data_helpers.py
@@ -1,0 +1,558 @@
+import numpy as np
+import pandas as pd
+from scipy.integrate import quad
+from pyswarm import pso
+from scipy.interpolate import PchipInterpolator
+from scipy.interpolate import interp1d
+import matplotlib.pyplot as plt
+from scipy.optimize import curve_fit
+import os
+import warnings
+import tkinter as tk
+from tkinter import messagebox
+from astropy.timeseries import LombScargle
+from scipy.signal import find_peaks, peak_widths, peak_prominences, find_peaks_cwt
+from scipy.integrate import quad
+from matplotlib.colors import Normalize
+from matplotlib.cm import ScalarMappable
+import inspect
+import math
+#..................................................... Properly prepare imput for fitting functions
+def _define_x_y_sy(x_data, y_data, y_err):
+    """
+    Prepare and validate x, y, and error data for model fitting.
+
+    This function processes the input data by ensuring that `x_data`, `y_data`,
+    and `y_err` are properly formatted as NumPy arrays. It also calculates error weights
+    (`y_err_weight`) based on the provided uncertainties. If no uncertainties are provided,
+    a default weight of 1 is assigned. Additionally, the function handles time binning
+    when `x_data` has one more element than `y_data`.
+
+    Parameters
+    ----------
+    x_data : array-like
+        The independent variable data (e.g., time or position data).
+    y_data : array-like
+        The dependent variable data (e.g., observed values).
+    y_err : array-like
+        The uncertainties (errors) associated with `y_data`. Can be a 1D array of errors
+        or a 2D array (e.g., lower and upper bounds).
+
+    Returns
+    -------
+    x_data : array-like
+        The processed independent variable data, possibly as time bin pairs.
+    y_err_weight : array-like
+        The calculated or default error weights for the fitting process.
+    """
+    x_data = np.array(x_data)
+    y_data = np.array(y_data)
+    y_err = np.array(y_err)
+    
+    # If uncertainties are provided, calculate the error weights
+    if np.sum(y_err) != 0:
+        if len(y_err.shape) == 2:
+            y_err_weight = np.array(y_err[1] + y_err[0])  # Sum upper and lower bounds if 2D
+        elif len(y_err.shape) == 1:
+            y_err_weight = np.array(y_err)  # Use errors directly if 1D
+    
+    # If no uncertainties are provided, set all weights to 1
+    if np.sum(y_err) == 0:
+        y_err_weight = np.ones(len(y_data))
+    
+    # If `x_data` has one more element than `y_data`, assume it's time bin edges and create pairs
+    if len(y_data) + 1 == len(x_data) and len(x_data.shape) == 1:
+        x_data = _time_pairs(x_data)
+        
+    return np.array(x_data), np.array(y_err_weight)
+
+
+#................................................... Easy way to manage values
+def _copy_fields(source_file, destination_file, params2=None):
+    """
+    Copy and update fields from a source file to a destination file, matching parameter names and updating values accordingly.
+
+    This function reads parameters and values from a `source_file`, compares them with the parameters in a `destination_file`,
+    and updates the destination file with values from the source where the parameter names match. If the destination file
+    does not exist, it can be created using `params2`, and machine parameters are loaded from the `source_file`.
+
+    Parameters
+    ----------
+    source_file : str
+        The path to the source file, which contains the original parameter names and values.
+    destination_file : str
+        The path to the destination file where the updated parameter values will be written.
+    params2 : list of str, optional
+        A list of parameter names to use when creating the destination file if it does not already exist.
+        If not provided, and the destination file is missing, a ValueError is raised.
+
+    Raises
+    ------
+    ValueError
+        If the destination file does not exist and `params2` is not provided, or if `params2` is empty.
+
+    Notes
+    -----
[…]
