+++ xraybinaryorbit/fitting/fit_spiral_orbit.py
@@ -1,0 +1,492 @@
+import numpy as np
+import pandas as pd
+from scipy.integrate import quad
+from pyswarm import pso
+from scipy.interpolate import PchipInterpolator
+from scipy.interpolate import interp1d
+import matplotlib.pyplot as plt
+from scipy.optimize import curve_fit
+import os
+import warnings
+import tkinter as tk
+from tkinter import messagebox
+from astropy.timeseries import LombScargle
+from scipy.signal import find_peaks, peak_widths, peak_prominences, find_peaks_cwt
+from scipy.integrate import quad
+from matplotlib.colors import Normalize
+from matplotlib.cm import ScalarMappable
+import inspect
+import math
+
+from ..helpers.data_helpers import _manage_parameters,_define_x_y_sy,_copy_fields, _load_values_to_interface, _manage_parameters,_load_bounds_to_interface, _manage_bounds
+
+from ..helpers.math_helpers import _gaussian,_time_pairs,_interpolate_pchip,_chi_squared_weighted,_chi_squared,_orbital_phase_to_time,_orbital_time_to_phase
+
+c = 299792458
+
+msun = (1.98847*10**30)*1000 #gr
+rsun_m = 696340*1000 #
+rsun_cm = 696340*1000*100 #cm
+
+kev_ams = 1.23984193
+
+na = 6.02214076*10**23/1.00797
+mu = 0.5
+mp = 1.67E-24
+
+# SPIRAL IN ORBIT #############################################################################
+def _spiral_orbit(x_data, iphase_orbit, semimajor_orbit, orbitalperiod, eccentricity, periapsis, inclination_orbit, Rstar, Mstar1, Mstar2, iphase_spiral, semimajor_spiral, b, omega, inclination_spiral, feature, units, method_, extended_binsize):
+    """
+    Simulate the Doppler shifts for a system that includes both a main orbital motion and a spiral structure. This
+    private function is used in fit_spiral_in_orbit_ps and fit_spiral_in_orbit_ls.
+
+    This function calculates the Doppler shifts in a system that has both an elliptical orbit (main orbit)
+    and an expanding spiral structure. The Doppler shifts are computed for both the main orbital motion
+    and the spiral structure, and the results are combined. The Doppler shifts can be returned in various
+    units, such as keV, seconds, or wavelength (angstroms). The function supports both "extended" and "discrete"
+    calculation methods.
+
+    Parameters
+    ----------
+    x_data : array-like
+        The input time data (or time bins) to compute the Doppler shifts for both the main orbit and spiral structure.
+    iphase_orbit : float
+        The initial phase of the main orbit, typically as a fraction of the orbital period or in radians.
+    semimajor_orbit : float
+        The semi-major axis of the main orbit, in solar radii or other relevant units.
+    orbitalperiod : float
+        The orbital period of the main system in days.
+    eccentricity : float
+        The eccentricity of the main orbit, ranging from 0 (circular) to 1 (parabolic).
+    periapsis : float
+        The argument of periapsis of the main orbit, describing the orientation of the orbit in the plane of motion.
+    inclination_orbit : float
+        The inclination of the main orbit in degrees, relative to the plane of the sky.
+    Rstar : float
+        The radius of the star, typically in solar radii.
+    Mstar1 : float
+        The mass of the primary object (main star), typically in solar masses.
+    Mstar2 : float
+        The mass of the secondary object, typically in solar masses.
+    iphase_spiral : float
+        The initial phase of the spiral structure, typically as a fraction of the orbital period or in radians.
[…]
