+++ xraybinaryorbit/timing/__init__.py
@@ -1,0 +1,2 @@
+from ..main import *
+from .timing_related import *
