+++ examples/example_fexxv_doppler_shifts_CenX3/bounds_spiral_orbit.txt
@@ -1,3 +1,3 @@
-iphase_orbit,semimajor_orbit,orbitalperiod,eccentricity,periapsis,inclination,Rstar,Mstar1,Mstar2,iphase_spiral,semimajor_spiral,b,omega,inclination_spiral,feature
+iphase_orbit,semimajor_orbit,orbitalperiod,eccentricity,periapsis,inclination,Rstar,Mstar1,Mstar2,iphase_spiral,semimajor_spiral,b,omega,inclination_spiral
 0.18,1.56,2.08,0.0,0.0,67.0,11.8,18.7,1.33,0.0,0.01,-0.1,1e-07,67.0,6.63
 0.25,1.6,2.1,1e-05,1e-05,77.0,12.2,22.77,1.45,1.0,100.0,0.0,0.1,77.0,6.7
