+++ examples/example_fexxv_doppler_shifts_CenX3/bounds_orbit.txt
@@ -1,3 +1,3 @@
 iphase,semimajor,orbitalperiod,eccentricity,periapsis,inclination,Rstar,Mstar1,Mstar2,wind_vel,feature
-0.0,0.01,0.05,0.0,0.0,68.0,11.8,18.7,1.33,0.0,6.63
-1.0,3.0,2.0,1e-05,1e-05,76.0,12.2,22.77,1.45,1.0,6.7
+0.18,1.5,2.087,0.0,0.0,67.0,11.8,18.7,1.33,0.0,6.63
+0.25,1.7,2.09,1e-05,1e-70,76.7,12.2,22.77,1.45,0.99,6.67

[…]
