+++ xraybinaryorbit/helpers/__init__.py
@@ -22,6 +22,8 @@ from .math_helpers import (
     _orbital_time_to_phase
 )
 
+
+
 # Define __all__ to make these functions available for import
 __all__ = [
     '_define_x_y_sy',
@@ -36,5 +38,6 @@ __all__ = [
     '_chi_squared_weighted',
     '_chi_squared',
     '_orbital_phase_to_time',
-    '_orbital_time_to_phase'
+    '_orbital_time_to_phase',
+
 ]
