+++ xraybinaryorbit/__init__.py
@@ -5,6 +5,7 @@ from .timing import *
 from .fitting import *
 from .theoretical import *
 
+
 from .helpers.data_helpers import (
     _define_x_y_sy,
     _copy_fields,
