+++ xraybinaryorbit/fitting/fit_nh.py
@@ -224,7 +224,6 @@ def fit_nh_ps(x_data, y_data, y_err=0, num_iterations=3, maxiter=200, swarmsize=
 
 
 #............................................Data prep.
-    _advise()
     parameter_names=["iphase", "semimajor", "orbitalperiod", "eccentricity", "periapsis" ,"inclination", "Rstar", "Mstar1", "Mstar2", "Mdot", "v_inf", "beta"]
     
     t = x_data
