### 2024-11-22 #7220 @warrickball
Thanks @matteobachetti and @renecotyfanboy! @xragua, that concludes a successful review—congratulations!

Let me now when you made your final changes and we'll start working through a post-review checklist. The main item for you will be to make a new tagged release and upload the archive to an archiving service like Zenodo or Figshare. There are multiple ways of doing it but a common one is to [set up a GitHub integration with Zenodo](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content). If you go down this route, I think you need to activate the integration *before* creating the release, so that it's picked up.

You can then let me know the Zenodo DOI and version number, which I'll add to the JOSS metadata.
