### 2024-11-24 #7220 @xragua
Dear @matteobachetti and @renecotyfanboy, Thank you so much for your review, guidance, and patience! 😊 My code has improved significantly, and I’ve really learned a great deal from this process thanks to you.

@warrickball, thank you so much for your work as editor. I will review and set up the integration with Zenodo and will write back as soon as I’m finished.
