+++ my-project/site/notebooks/example_fexxv_doppler_shifts_CenX3/bounds_spiral_in_orbit.txt
@@ -1,2 +1,0 @@
-iphase,semimajor,orbitalperiod,eccentricity,periapsis,inclination,iphase_spiral,semimajor_spiral,b,omega,inclination_spiral,Rstar,Mstar1,Mstar2,wind_vel,feature
-0.2,3.0,2.1,0.3,120.0,78.0,0.9,0.3,0.01,0.001,67.0,23.0,12.0,1.4,1200.0,6.4
