+++ my-project/docs/index.md
@@ -47,7 +47,7 @@ So, dive in our **EXAMPLES** where we show some interesting uses cases, theoreti
 ### General Usage
 
 The software provides a user-friendly interface for managing the various parameters that influence orbital modulations. Upon first use, the user inputs parameters through a form, which are then saved in a file within the running directory. This file is automatically loaded in future runs, eliminating the need to re-enter all parameters.
- If the file is absent, the form will reappear for new inputs. Alternatively, setting `load_directly=True` will bypass the form and run the code using previously saved parameters (only if the file exists).
+ If the file is absent, the form will reappear for new inputs. By setting `load_directly=True` will bypass the form and run the code using previously saved parameters (only if the file exists).
  Alternatively, the parameters can be provided as lists in the following format: parameter_list=[parameters] for the theoretical functions or as bound_list = [lower_bounds], [upper_bounds] for the fitting functions.
  
 
