+++ paper/paper.md
@@ -49,7 +49,7 @@ Although these orbital modulations are conceptually simple, they are challenging
 
 Historically, the primary limitation in this type of analysis has been the lack of resolution for detailed phase-resolved observations. However, upcoming high-resolution missions like XRISM [@2022arXiv220205399X] and New Athena [@2016SPIE.9905E.2FB] promise to significantly improve the quality of these analyses. In addition to better resolution, advances in computational power have been crucial. Many of these tools have already been successfully applied to studies using XMM-Newton and Chandra data, enabling analyses that were previously impossible. [@2022MNRAS.512.304S;@2021MNRAS.501.5892S;@newcenx3]
 
-**xraybinaryorbit** is a user-friendly Python package designed to simplify the analysis of orbital modulations in X-ray binaries. The package is organized into three groups of functions: theoretical, fitting, and timing. It features an intuitive interface for managing the various parameters that influence orbital modulations. Upon first use, the user inputs parameters via a form; these parameters are then saved to a file within the working directory. This file is automatically loaded in future sessions, eliminating the need to re-enter parameters each time. Alternatively, the parameters can be provided as lists in the following format: parameter_list=[parameters] for the theoretical functions or as bound_list = [lower_bounds], [upper_bounds] for the fitting functions.
+**xraybinaryorbit** is a user-friendly Python package designed to simplify the analysis of orbital modulations in X-ray binaries. The package is organized into three groups of functions: theoretical, fitting, and timing. It features an intuitive interface for managing the various parameters that influence orbital modulations. Upon first use, the user inputs parameters via a form; these parameters are then saved to a file within the working directory. This file is automatically loaded in future sessions, eliminating the need to re-enter parameters each time. 
 
 The software offers two fitting methods: least squares (LS) and particle swarm optimization (PSO) [@10.1162/EVCO_r_00180, @pyswarms]. The LS method is faster but may fail to converge in some cases, whereas PSO is more robust but computationally intensive. Some results obtained with this package are shown in Figure 1.
 
