### 2024-11-25 #7220 @xragua
Dear @warrickball , I have created a new release, which is now available on Zenodo: DOI 10.5281/zenodo.14216915, version 1.0.00.
Please let me know if there are any further instructions or actions required.

Thank you very much!
