### 2024-11-26 #7220 @warrickball
Thanks @xragua! You [previously wrote](https://github.com/openjournals/joss-reviews/issues/7220#issuecomment-2484465384)

> Thank you very much for all the advices! This is quite new to me and I really appreciate all the guidance :) I will write again when I get to finish with the new functionallity.

Is the new functionality finished?
