### 2024-11-26 #7220 @warrickball
Regarding the archive, note also that title and author list should match the paper and the licence of the archive (currently CC-BY-4.0) should also match the repo (MIT). You should be able to change these details without creating a new archive.
