### 2024-11-27 #7220 @xragua
Dear @warrickball, I think I made the changes correctly and now the authors and title match the paper's, and also changed the license to the MIT on the archive, (the new functionality was also finished).
I will be waiting for further instructions, thanks a lot!
