### 2024-12-01 #7220 @warrickball
Hi @xragua, I've just started having a look at the paper and have a few comments, starting with a broad one. Most of the summary (lines 8–34) contains no references at all, which I think needs correcting. I'm not expecting anything hugely detailed but references to a few key review papers, books or high-impact results over the years would be appropriate.

Note that there's also a mix of XRAYBINARYORBIT (e.g. [the repo README](https://github.com/xragua/xraybinaryorbit))  and xraybinaryorbit (e.g. [the docs](https://xragua.github.io/xraybinaryorbit/)) as the software's name. Because we don't really support things like small caps, as other journals do, we tend towards having all lowercase names. Note that when I make the editorial corrections (when I put my assistant editor-in-chief hat on), I'll remove all the boldface text anyway.

As a smaller point, Fig. 1 has embedded text that I won't be able to propose corrections to via pull request. Note that "Particle swarm" should be "particle swarm", "Real Ionization" should be "Real ionization" and you've written "trayectory" instead of "trajectory"

I'd also suggest moving the _Statement of Need_ to follow the _Summary_. This shouldn't affect the flow of the text, and I think the last two paragraphs of the _Summary_ could be places at the end of the _Statement of Need_. I.e., in effect, you could move the _Statement of Need_ to what is [currently ](https://github.com/openjournals/joss-reviews/issues/7220#issuecomment-2510245807) the start of page 2 but maybe even line 35.

Finally, if any of your co-authors have ORCIDs, you should add them, but I realise not all researchers have them.
