### 2024-12-02 6569318
improve nh calculations
