### 2024-12-02 515bc19
improve nh calculations
