+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'XRAYBINARYORBIT: A Python Package for Analyzing Orbital Modulations in X-ray Binaries'
+title: 'xraybinaryorbit: A Python Package for Analyzing Orbital Modulations in X-ray Binaries'
 tags:
   - Python
   - x-ray astronomy
@@ -35,13 +35,13 @@ bibliography: paper.bib
 
 # Summary
 
-X-ray astronomy is a young discipline, spanning no more than a few decades. The Earth's atmosphere is opaque to this type of radiation, so observations in this part of the spectrum were delayed until the space era began, when rocket launchers carrying X-ray telescopes revealed the universe from a brand-new point of view.
+X-ray astronomy is a young discipline, spanning no more than a few decades. The Earth's atmosphere is opaque to this type of radiation, so observations in this part of the spectrum were delayed until the space era began, when rocket launchers carrying X-ray telescopes revealed the universe from a brand-new point of view (see [@CambridgeXrayHistory] for a nice overview).
 
-X-ray binaries are tremendously interesting systems. In these extreme environments, a compact object (either a neutron star, a black hole or a white dwarf) draws in matter from a companion star, producing X-ray radiation. These systems offer a unique window into extreme physics, from the effects of strong gravity and relativistic jets to the presence of intense magnetic fields.
+X-ray binaries are tremendously interesting systems. In these extreme environments, a compact object (either a neutron star, a black hole or a white dwarf) draws in matter from a companion star, producing X-ray radiation. These systems offer a unique window into extreme physics, from the effects of strong gravity and relativistic jets to the presence of intense magnetic fields [@Carroll_Ostlie_2017;@1983bhwd.book.S].
 
 Orbital modulations are observed in nearly all X-ray binary systems. These variations arise from the orbital motions within the system, driven by the relative velocities of the two stars and/or their changing configurations with respect to each other and the observer. 
 
-For instance, we might detect slight oscillating shifts in the center of emission lines and/or small variations in a neutron star's spin period. This phenomenon is likely caused by the Doppler effect.Our code translates this data into orbital parameters that explain these Doppler shifts. Although the process seems simple, it becomes complex due to factors like inclination, eccentricity, periapsis, and the masses of the stars involved. If the eccentricity is greater than zero, the velocity around the orbit is not constant, making the analysis even more intricate.
+For instance, we might detect slight oscillating shifts in the center of emission lines and/or small variations in a neutron star's spin period. This phenomenon is likely caused by the Doppler effect. Our code translates this data into orbital parameters that explain these Doppler shifts. Although the process seems simple, it becomes complex due to factors like inclination, eccentricity, periapsis, and the masses of the stars involved. If the eccentricity is greater than zero, the velocity around the orbit is not constant, making the analysis even more intricate.
 
 When considering stellar wind (the matter accreted by compact objects), various combinations of these factors result in orbital modulations. Eccentricity causes changes in wind density throughout the orbit, affecting the amount of accreted matter. Both eccentricity and inclination influence the absorption column faced by the emitted radiation and the ionization of the stellar wind, which also varies with orbital phase.
 
@@ -49,7 +49,13 @@ Although these orbital modulations are conceptually simple, they are challenging
 
[…]
