+++ paper/paper.bib
@@ -330,3 +330,26 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2024A&A...690A.317D},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+
+@misc{CambridgeXrayHistory,
+  author = {{University of Cambridge}},
+  title = {History of X-ray astronomy},
+  year = {n.d.},
+  url = {https://www-xray.ast.cam.ac.uk/xray_introduction/History.html},
+  note = {Accessed: 2024-12-02}
+}
+
+@inbook{Carroll_Ostlie_2017, place={Cambridge}, title={General Relativity and Black Holes}, booktitle={An Introduction to Modern Astrophysics}, publisher={Cambridge University Press}, author={Carroll, Bradley W. and Ostlie, Dale A.}, year={2017}, pages={609–652}} 
+
+
+
+@BOOK{1983bhwd.book.S,
+       author = {{Shapiro}, Stuart L. and {Teukolsky}, Saul A.},
+        title = "{Black holes, white dwarfs and neutron stars. The physics of compact objects}",
+         year = 1983,
+          doi = {10.1002/9783527617661},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/1983bhwd.book.....S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
