+++ paper/paper.bib
@@ -353,3 +353,14 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+
+@book{lewin1997xray,
+  title     = {X-ray Binaries},
+  author    = {Lewin, Walter H. G. and Van Paradijs, Jan and Van Den Heuvel, Edward P. J.},
+  year      = {1997},
+  publisher = {Cambridge University Press},
+  series    = {Cambridge Astrophysics Series},
+  volume    = {26},
+  isbn      = {9780521599344},
+}
+
