### 2024-12-02 #7220 @xragua
Hello @warrickball, I think I addressed all the points mentioned above, but I have one doubt, one of the references I added is in t Cambridge web page in which a nice (and quite informal) summary of X-ray astronomy is presented (https://www-xray.ast.cam.ac. uk/xray_introduction/History.html) and I'm not sure if is adequate. Let me know what you think, I will be waiting for more indications, thank you very much!!
