### 2024-12-03 e297aff
update nh calc
