### 2024-12-05 #7220 @warrickball
Thanks for the additional references.

> ... one of the references I added is in t Cambridge web page ...

Unless somehow authoritative, I think original, peer-reviewed sources are much better. In this particular case, I think the relevant reference for Giacconi et al.'s pioneering observations would be [Giacconi et al. (1962)](https://ui.adsabs.harvard.edu/abs/1962PhRvL...9..439G).
