+++ paper/paper.bib
@@ -364,3 +364,18 @@ archivePrefix = {arXiv},
   isbn      = {9780521599344},
 }
 
+
+@ARTICLE{1962PhRvL.9.439G,
+       author = {{Giacconi}, Riccardo and {Gursky}, Herbert and {Paolini}, Frank R. and {Rossi}, Bruno B.},
+        title = "{Evidence for x Rays From Sources Outside the Solar System}",
+      journal = {\prl},
+         year = 1962,
+        month = dec,
+       volume = {9},
+       number = {11},
+        pages = {439-443},
+          doi = {10.1103/PhysRevLett.9.439},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/1962PhRvL...9..439G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
