### 2024-12-09 #7220 @warrickball
Thanks for these changes @xragua, we're getting there! As another small point, the section title "Science Behind" doesn't make sense on its own. Perhaps you want something like "Background theory", "Scientific background" or anything else along those lines.

More substantially regarding the image again, the fourth caption should presumably be "Ionization maps of the **orbital** plane...". We also tend not to have so many images in JOSS papers, so I'd suggest removing the second (the fit to the spin period evolution), because it does the least to demonstrate the features of xraybinaryorbit.
You should then split each of the remaining panels into their own figure, with the caption in the text rather than embedded in the image. (The images are quite wide so I expect the won't take up that much vertical space on the page.)

Finally, you'll need to refer to all three figures in the text. I.e., the last sentence in the *Statement of Need* will have to become

> Some results obtained with this package are shown in **Figures 1, 2 and 3**.

Splitting up the figures, getting the text into the captions and removing the second fit (the basic curve fit) are the most important things. I can propose changes to the text via a pull request after that, if you have any issues yourself.
