+++ paper/paper.md
@@ -96,11 +96,11 @@ where r is the orbital radius, $\omega$ is angular velocity, $\i$ is inclination
 
 # Examples of Package Performance
 
-![Reconstructed trajectory of Fe xxv emitting plasma for Cen X-3, deduced by the observed Doppler shifts of the emission line center.](trajectory_and_shifts.png){#sylt width="100%"}
+![Reconstructed trajectory of Fe XXV emitting plasma in Cen X-3, derived from the observed Doppler shifts of the emission line center.](trajectory_and_shifts.png){#sylt width="100%"}
 
-![Ionization map of the stellar wind on the orbital plane for 4U 0114+65 for the different luminosities observed.](new_iopn_map.png){#sylt width="100%"}
+![Ionization map of the stellar wind in the orbital plane of 4U 0114+65, showing variations in different observed luminosities.](new_iopn_map.png){#sylt width="100%"}
 
-![Theoretical local absoprption column for a hipothetical system.](NH_through_the_orbit.png){#sylt width="100%"}
+![Theoretical local absorption column for a hypothetical system, as calculated through the orbit.](NH_through_the_orbit.png){#sylt width="100%"}
 
 
 # Acknowledgements
