### 2024-12-14 #7220 @warrickball
This looks good! I've just opened [a pull request](https://github.com/xragua/xraybinaryorbit/pull/22) with a first set of editorial changes to the paper.

I'll have another look at the submission details when I test run the publication pipeline but this one should already catch most issues.
