+++ paper/paper.bib
@@ -104,7 +104,6 @@ archivePrefix = {arXiv},
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@book{Frank_King_Raine_2002, place={Cambridge}, edition={3}, title={Accretion Power in Astrophysics}, publisher={Cambridge University Press}, author={Frank, Juhan and King, Andrew and Raine, Derek}, year={2002}} <div></div>
 
 
 @ARTICLE{2017PhRvL.119p1101A,
