### 2024-12-17 #7220 @warrickball
Hi @xragua — I suspect there's a small issue in the names in the metadata. Specifically, the names are currently being interpreted as (boldface indicates **surname**):

1. Graciela **Sanjurjo-Ferrín**
2. Jessica Planelles **Villalva**
3. Jose Miguel **Torrejón**
4. Jose Joaquín **Rodes-Roca**

I think (from the [ORCID](https://orcid.org/0009-0007-9988-0202)) author 2 should be *Jessica **Planelles Villalva***.

If I've understood correctly, you can indicate the specific components of [names](https://joss.readthedocs.io/en/latest/paper.html#names) by replacing
```
  - name: Jessica Planelles Villalva
```
with
```
  - given-names: Jessica
  - surname: Planelles Villalva
```
I'll leave this to you to change in case you need to make similar corrections to the other names but I *think* they've been automatically interpreted correctly. If you let me know when it's fixed, I'll do another test run of the publication pipeline.

Thanks for your patience!
