+++ paper/paper.md
@@ -9,19 +9,23 @@ tags:
   - neutron stars
   - black holes
 authors:
-  - name: Graciela Sanjurjo-Ferrín
+  - given-names: Graciela
+    surname: Sanjurjo-Ferrín
     orcid: 0009-0001-0378-7879
     equal-contrib: true
     affiliation: 1
-  - name: Jessica Planelles Villalva
+  - given-names: Jessica 
+    surname: Planelles Villalva
     orcid: 0009-0007-9988-0202
     equal-contrib: true
     affiliation: 1
-  - name: Jose Miguel Torrejón
+  - given-names: Jose Miguel 
+    surname: Torrejón
     orcid: 0000-0002-5967-5163
     equal-contrib: true
     affiliation: 1
-  - name: Jose Joaquín Rodes-Roca
+  - given-names: Jose Joaquín 
+    surname: Rodes-Roca
     orcid: 0000-0003-4363-8138
     equal-contrib: true
     affiliation: 1
