### 2024-12-18 #7220 @xragua
Hi @warrickball, I corrected the names in the paper (they were all correct except for Jessica's), I hope that now the metadata looks good ( I don't know where should I check it...) please, let me know if I made any mistakes or It needs further corrections.

Thank you so much for the indications!
