### 2024-12-18 #7220 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
