### 2024-12-18 #7220 @warrickball
Thanks once again to @matteobachetti & @renecotyfanboy for reviewing this submission! JOSS simply wouldn't be possible without its community of volunteers.
