### 2024-12-18 #7220 @xragua
@matteobachetti, @renecotyfanboy, @warrickball, we got a paper! 🎉🎉🎉 Thank you so much for your efforts and collaboration. It was a real pleasure working with you, all of you were incredibly helpful and patient. Wishing you a wonderful holiday season, and I hope we get the chance to collaborate again in the future!
